### 2025-05-02 #8150 @editorialbot
[PRE REVIEW]: JAX-bandflux: differentiable supernovae SALT modelling for cosmological analysis on GPUs
**Submitting author:** @samleeney (<a href="http://orcid.org/0000-0003-4366-1119">Samuel Alan Kossoff Leeney</a>)
**Repository:** https://github.com/samleeney/JAX-bandflux
**Branch with paper.md** (empty if default branch):
**Version:** v0.3.1
