### 2025-05-02 #8150 @warrickball
Hi @samleeney! Thanks for your submission and sorry for the long wait to start the pre-review process. All the suitable JOSS editors are currently working at capacity so I'm going to "waitlist" this review until an editor with the relevant expertise is available to take it on. The AASS track also currently has a substantial backlog, so it might be several months before an editor picks this up. Thanks for your patience!
