+++ tests/test_dust_extinction.py
@@ -1,0 +1,315 @@
+"""Tests for dust extinction laws in JAX-bandflux."""
+
+import numpy as np
+import jax
+import jax.numpy as jnp
+import pytest
+import sncosmo
+from jax_supernovae import dust
+from jax_supernovae.bandpasses import Bandpass
+from jax_supernovae.salt3 import optimized_salt3_bandflux, precompute_bandflux_bridge
+
+# Enable float64 precision
+jax.config.update("jax_enable_x64", True)
+
+
+def test_dust_extinction_laws():
+    """Test that JAX dust extinction laws match sncosmo's implementation."""
+    # Test wavelength grid
+    wave = np.linspace(3000, 10000, 100)
+    
+    # Test parameters
+    ebv = 0.1
+    r_v = 3.1
+    
+    # Test CCM89 dust law
+    ccm89_jax = dust.ccm89_extinction(wave, ebv, r_v)
+    ccm89_sncosmo = sncosmo.CCM89Dust()
+    ccm89_sncosmo.parameters = [ebv, r_v]
+    ccm89_sncosmo_flux = ccm89_sncosmo.propagate(wave, np.ones_like(wave))
+    ccm89_sncosmo_extinction = -2.5 * np.log10(ccm89_sncosmo_flux)
+    
+    # Check that the results are close
+    np.testing.assert_allclose(ccm89_jax, ccm89_sncosmo_extinction, rtol=1e-3)
+    
+    # Test OD94 dust law
+    od94_jax = dust.od94_extinction(wave, ebv, r_v)
+    od94_sncosmo = sncosmo.OD94Dust()
+    od94_sncosmo.parameters = [ebv, r_v]
+    od94_sncosmo_flux = od94_sncosmo.propagate(wave, np.ones_like(wave))
+    od94_sncosmo_extinction = -2.5 * np.log10(od94_sncosmo_flux)
+    
+    # Check that the results are close
+    np.testing.assert_allclose(od94_jax, od94_sncosmo_extinction, rtol=1e-3)
+    
+    # Test F99 dust law
+    f99_jax = dust.f99_extinction(wave, ebv, r_v)
+    f99_sncosmo = sncosmo.F99Dust(r_v=r_v)
+    f99_sncosmo.parameters = [ebv]
+    f99_sncosmo_flux = f99_sncosmo.propagate(wave, np.ones_like(wave))
+    f99_sncosmo_extinction = -2.5 * np.log10(f99_sncosmo_flux)
+    
+    # Check that the results are close with a higher tolerance for F99
+    # F99 is more complex and our implementation is based on interpolation
+    np.testing.assert_allclose(f99_jax, f99_sncosmo_extinction, rtol=6e-2)
+
+
+def test_dust_extinction_laws_uv():
+    """Test dust extinction laws in the UV range where issues were occurring."""
+    # Test wavelength grid focusing on UV range
+    wave = np.linspace(2000, 3500, 100)
+    
+    # Test parameters
+    ebv = 0.1
+    r_v = 3.1
+    
+    # Test CCM89 dust law
+    ccm89_jax = dust.ccm89_extinction(wave, ebv, r_v)
+    ccm89_sncosmo = sncosmo.CCM89Dust()
+    ccm89_sncosmo.parameters = [ebv, r_v]
+    ccm89_sncosmo_flux = ccm89_sncosmo.propagate(wave, np.ones_like(wave))
+    ccm89_sncosmo_extinction = -2.5 * np.log10(ccm89_sncosmo_flux)
+    
+    # Check that the results are close
+    np.testing.assert_allclose(ccm89_jax, ccm89_sncosmo_extinction, rtol=1e-2)
+    
[…]
