+++ jax_supernovae/salt3.py
@@ -11,6 +11,7 @@ from jax_supernovae.bandpasses import HC_ERG_AA, MODEL_BANDFLUX_SPACING
 from functools import partial
 from jax import vmap
 import importlib.resources
+from jax_supernovae import dust
 
 # Enable float64 precision
 jax.config.update("jax_enable_x64", True)
@@ -125,7 +126,7 @@ def kernval(x):
     
     # Use where to select the appropriate result
     result = jnp.where(x <= 1, case1,
-                      jnp.where(x < 2, case2, 0.0))
+                       jnp.where(x < 2, case2, 0.0))
     
     return result
 
@@ -424,6 +425,10 @@ def salt3_bandflux(phase, bandpass, params, zp=None, zpsys=None):
         Bandpass to calculate flux through.
     params : dict
         Model parameters including z, t0, x0, x1, c.
+        Optional dust parameters:
+        - 'dust_type': int, dust law index (0=ccm89, 1=od94, 2=f99)
+        - 'ebv': float, E(B-V) value
+        - 'r_v': float, R_V value (default: 3.1)
     zp : float or None, optional
         Zero point for flux. If None, no scaling is applied.
     zpsys : str, optional
@@ -471,6 +476,30 @@ def salt3_bandflux(phase, bandpass, params, zp=None, zpsys=None):
     
     # Calculate rest-frame flux for all phases and wavelengths at once
     rest_flux = x0 * (m0 + x1 * m1) * 10**(-0.4 * cl[None, :] * c) * a
+    
+    # Apply dust extinction if parameters are provided
+    has_dust = 'dust_type' in params and 'ebv' in params
+    if has_dust:
+        dust_type_idx = params['dust_type']
+        ebv = params['ebv']
+        r_v = params.get('r_v', 3.1)  # Default R_V = 3.1 if not specified
+        
+        # Get the appropriate dust law function based on the index
+        if dust_type_idx == 0:
+            dust_law = dust.ccm89_extinction
+        elif dust_type_idx == 1:
+            dust_law = dust.od94_extinction
+        elif dust_type_idx == 2:
+            dust_law = dust.f99_extinction
+        else:
+            # Default to CCM89
+            dust_law = dust.ccm89_extinction
+        
+        # Calculate extinction for each wavelength
+        extinction = dust_law(restwave, ebv, r_v)
+        
+        # Apply extinction to rest-frame flux
+        rest_flux = dust.apply_extinction(rest_flux, extinction[None, :])
     
     # Integrate flux through bandpass using trapezoidal rule
     # Multiply by wave and transmission before integration
@@ -566,6 +595,10 @@ def optimized_salt3_bandflux(phase, wave, dwave, trans, params, zp=None, zpsys=N
         Transmission values on the wavelength grid
     params : dict
         Dictionary containing model parameters: 'z', 't0', 'x0', 'x1', 'c'
+        Optional dust parameters:
+        - 'dust_type': int, dust law index (0=ccm89, 1=od94, 2=f99)
+        - 'ebv': float, E(B-V) value
+        - 'r_v': float, R_V value (default: 3.1)
     zp : float or None, optional
         Zero point for flux scaling. If provided, zpsys must also be given
     zpsys : str or None, optional
@@ -604,6 +637,35 @@ def optimized_salt3_bandflux(phase, wave, dwave, trans, params, zp=None, zpsys=N
 
     # Compute rest-frame flux including the colour law effect.
     rest_flux = x0 * (m0 + x1 * m1) * 10**(-0.4 * cl[None, :] * c) * a
+    
+    # Apply dust extinction if parameters are provided
[…]
