### 2025-06-05 34eaf31
Update JAX and BlackJAX dependencies for GPU compatibility
- Update requirements.txt to use jax[cuda12] for latest CUDA support
- Migrate from blackjax.ns.adaptive.nss to blackjax.nss API
- Update nested sampling parameter names: num_mcmc_steps -> num_inner_steps
- Fix state access patterns: state.sampler_state.logZ -> state.logZ
- Update algorithm initialization to new BlackJAX interface
- All dust extinction tests passing (7/7)
- Core functionality verified with updated syntax

🤖 Generated with [Claude Code](https://claude.ai/code)

Co-Authored-By: Claude <noreply@anthropic.com>
