+++ examples/ns.py
@@ -250,11 +250,11 @@ num_mcmc_steps = n_params_total * NS_SETTINGS['num_mcmc_steps_multiplier']
 
 # Initialize nested sampling algorithm
 print("Setting up nested sampling algorithm...")
-algo = blackjax.ns.adaptive.nss(
+algo = blackjax.nss(
     logprior_fn=logprior,
     loglikelihood_fn=compute_single_loglikelihood,
-    n_delete=NS_SETTINGS['n_delete'],
-    num_mcmc_steps=num_mcmc_steps,
+    num_inner_steps=num_mcmc_steps,
+    num_delete=NS_SETTINGS['n_delete'],
 )
 
 # Initialize random key and particles
@@ -265,7 +265,7 @@ initial_particles = sample_from_priors(init_key, NS_SETTINGS['n_live'])
 print("Initial particles generated, shape: ", initial_particles.shape)
 
 # Initialize state
-state = algo.init(initial_particles, compute_single_loglikelihood)
+state = algo.init(initial_particles)
 
 # Define one_step function with JIT
 @jax.jit
@@ -294,7 +294,7 @@ def one_step(carry, xs):
 dead = []
 print("Running nested sampling...")
 with tqdm.tqdm(desc="Dead points", unit=" dead points") as pbar:
-    while (not state.sampler_state.logZ_live - state.sampler_state.logZ < -3):
+    while (not state.logZ_live - state.logZ < -3):
         (state, rng_key), dead_info = one_step((state, rng_key), None)
         dead.append(dead_info)
         pbar.update(NS_SETTINGS['n_delete'])
@@ -308,7 +308,7 @@ dead = jax.tree_map(lambda *args: jnp.concatenate(args), *dead)
 logw = log_weights(rng_key, dead)
 logZs = jax.scipy.special.logsumexp(logw, axis=0)
 
-print(f"Runtime evidence: {state.sampler_state.logZ:.2f}")
+print(f"Runtime evidence: {state.logZ:.2f}")
 print(f"Estimated evidence: {logZs.mean():.2f} ± {logZs.std():.2f}")
 
 # Save chains using the utility function
