+++ requirements.txt
@@ -1,5 +1,4 @@
 numpy>=1.24.0
-jax>=0.5.3
 jaxlib==0.4.20
 sncosmo>=2.9.0
 pytest>=7.0.0 
