### 2025-06-05 3e683c2
Update dependencies and version to 0.3.2
- Update JAX to use cuda12 variant for GPU support
- Update BlackJAX to stable release (>=1.3.0)
- Add distrax as a core dependency
- Update installation instructions in README and CLAUDE.md
- Bump version to 0.3.2
