+++ requirements.txt
@@ -1,11 +1,12 @@
 numpy>=1.24.0
-jaxlib==0.4.20
+jax[cuda12]==0.4.35
+jaxlib==0.4.35
 sncosmo>=2.9.0
 pytest>=7.0.0 
 pyyaml>=6.0.0
 matplotlib
 tqdm
 anesthetic
-git+https://github.com/handley-lab/blackjax@proposal
+blackjax>=1.3.0
 requests
 distrax

[…]
