### 2025-06-05 902cfed
Fix BlackJAX dependency to use handley-lab fork
- Update to use git+https://github.com/handley-lab/blackjax.git
- Update installation instructions in CLAUDE.md
- Add installation section to README.rst
- All tests passing (dust extinction, core functionality)
