+++ setup.py
@@ -2,7 +2,7 @@ from setuptools import setup, find_packages
 
 setup(
     name="jax-bandflux",
-    version="0.3.2",
+    version="0.3.3",
     author="Samuel Alan Kossoff Leeney",
     author_email="sakl2@cam.ac.uk",
     description="A JAX-based package for calculating supernovae Bandfluxes",
@@ -38,7 +38,7 @@ setup(
         'tqdm',
         'anesthetic',
         'requests',
-        'blackjax>=1.3.0',
+        'blackjax @ git+https://github.com/handley-lab/blackjax.git',
         'distrax'
     ],
     extras_require={
