### 2025-06-08 ccbea68
Update examples and utils for new BlackJAX API
- Update nested sampling examples to use new BlackJAX API
- Change state.sampler_state.logZ to state.logZ
- Update jax.tree_map to jax.tree.map
- Update dead_info field names (logL to loglikelihood)
- Add device debugging prints in ns.py
- Remove JAX version constraints in requirements.txt

All examples now compatible with handley-lab BlackJAX fork
