+++ parameter_statistics.txt
@@ -2,8 +2,8 @@ Parameter Statistics:
 --------------------------------------------------
 Parameter               Mean         Std Dev
 --------------------------------------------------
-t0              58606.501734        0.622200
-log_x0             -2.968749        0.023750
-x1                  1.406445        0.684898
-c                   0.054883        0.066762
+t0              58606.506972        0.609268
+log_x0             -2.967710        0.022792
+x1                  1.400588        0.713620
+c                   0.055117        0.064023
 --------------------------------------------------
