+++ requirements.txt
@@ -1,6 +1,5 @@
 numpy>=1.24.0
-jax[cuda12]==0.4.35
-jaxlib==0.4.35
+jax[cuda12]
 sncosmo>=2.9.0
 pytest>=7.0.0 
 pyyaml>=6.0.0
