### 2025-08-06 07f36ce
Add transmission shift implementation for bandpass filters
- Modified bandpasses.py to support transmission wavelength shifts
- Updated salt3.py with shift-aware flux calculations
- Added comprehensive tests in test_transmission_shifts.py
- Created example nested sampling with shift parameters
