+++ examples/ns_with_shifts.py
@@ -1,0 +1,454 @@
+"""
+# Nested Sampling with JAX-bandflux including Transmission Shifts
+
+This script demonstrates how to run the nested sampling procedure for supernovae SALT model fitting using the JAX-bandflux package with transmission shift parameters for each filter.
+
+For more examples and the complete codebase, visit the [JAX-bandflux GitHub repository](https://github.com/samleeney/JAX-bandflux).
+"""
+
+import distrax
+import jax
+import jax.numpy as jnp
+import numpy as np
+import tqdm
+import blackjax
+import os
+from blackjax.ns.utils import log_weights
+from jax_supernovae.salt3 import (
+    optimized_salt3_multiband_flux,
+)
+from jax_supernovae.bandpasses import register_bandpass, get_bandpass, register_all_bandpasses
+from jax_supernovae.utils import save_chains_dead_birth
+from jax_supernovae.data import load_and_process_data
+import matplotlib.pyplot as plt
+from anesthetic import read_chains, make_2d_axes
+
+# Option to fit sigma or not.
+# When fit_sigma is True, the code will include an extra free parameter (log_sigma)
+# so that the effective flux error becomes sigma * flux_err with sigma = 10**(log_sigma).
+# When fit_sigma is False, the code will use the measured flux_err from the file.
+fit_sigma = False
+
+# Settings that were previously in YAML
+fix_z = True
+
+NS_SETTINGS = {
+    'n_delete': 1,
+    'n_live': 125,
+    'num_mcmc_steps_multiplier': 5
+}
+
+# Prior bounds (added log_sigma bound and shift bounds)
+PRIOR_BOUNDS = {
+    'z': {'min': 0.001, 'max': 0.2},
+    't0': {'min': 58000.0, 'max': 59000.0},
+    'x0': {'min': -5.0, 'max': -2.6},
+    'x1': {'min': -4.0, 'max': 4.0},
+    'c': {'min': -0.3, 'max': 0.3},
+    'log_sigma': {'min': -3.0, 'max': 1.0},
+    'shift': {'min': -50.0, 'max': 50.0}  # ±50 Å shift bounds
+}
+
+# Enable float64 precision
+jax.config.update("jax_enable_x64", True)
+
+# Load and process data
+times, fluxes, fluxerrs, zps, band_indices, bridges, fixed_z = load_and_process_data('19dwz', data_dir='data', fix_z=fix_z)
+
+# Get unique bands and their count
+unique_bands = np.unique(band_indices)
+n_bands = len(unique_bands)
+
+# Define parameter bounds and priors
+if fix_z:
+    param_bounds = {
+        't0': (PRIOR_BOUNDS['t0']['min'], PRIOR_BOUNDS['t0']['max']),
+        'x0': (PRIOR_BOUNDS['x0']['min'], PRIOR_BOUNDS['x0']['max']),
+        'x1': (PRIOR_BOUNDS['x1']['min'], PRIOR_BOUNDS['x1']['max']),
+        'c': (PRIOR_BOUNDS['c']['min'], PRIOR_BOUNDS['c']['max'])
+    }
+    # Add shift parameters for each band
+    for i in range(n_bands):
+        param_bounds[f'shift_{i}'] = (PRIOR_BOUNDS['shift']['min'], PRIOR_BOUNDS['shift']['max'])
+    
+    if fit_sigma:
+        param_bounds['log_sigma'] = (PRIOR_BOUNDS['log_sigma']['min'], PRIOR_BOUNDS['log_sigma']['max'])
+    
+    # Create prior distributions without z
+    prior_dists = {
+        't0': distrax.Uniform(low=param_bounds['t0'][0], high=param_bounds['t0'][1]),
+        'x0': distrax.Uniform(low=param_bounds['x0'][0], high=param_bounds['x0'][1]),
[…]
