+++ jax_supernovae/salt3.py
@@ -544,14 +544,53 @@ def precompute_bandflux_bridge(bandpass):
         - 'wave': the integration wavelength grid
         - 'dwave': spacing between grid points
         - 'trans': the transmission values computed on the grid
+        - 'wave_original': original wavelength array for shift interpolation
+        - 'trans_original': original transmission array
     """
     wave = bandpass.integration_wave
     dwave = bandpass.integration_spacing
     trans = bandpass(wave)
-    return {'wave': wave, 'dwave': dwave, 'trans': trans}
+    
+    # Store original arrays for shift interpolation
+    return {
+        'wave': wave, 
+        'dwave': dwave, 
+        'trans': trans,
+        'wave_original': bandpass.wave,
+        'trans_original': bandpass.trans
+    }
+
+@jax.jit
+def compute_shifted_transmission(wave, wave_original, trans_original, shift):
+    """Compute transmission values with wavelength shift.
+    
+    Parameters
+    ----------
+    wave : array
+        Wavelengths at which to evaluate transmission
+    wave_original : array
+        Original wavelength array from bandpass
+    trans_original : array
+        Original transmission array from bandpass
+    shift : float or array
+        Wavelength shift(s) to apply
+        
+    Returns
+    -------
+    array
+        Shifted transmission values
+    """
+    # Apply shift - if shift is callable, it should be evaluated outside JIT
+    effective_wave = wave - shift
+    
+    # Use existing interp function from utils
+    from jax_supernovae.utils import interp
+    return interp(effective_wave, wave_original, trans_original)
 
 @partial(jax.jit, static_argnames=['zpsys'])
-def optimized_salt3_bandflux(phase, wave, dwave, trans, params, zp=None, zpsys=None):
+def optimized_salt3_bandflux(phase, wave, dwave, trans, params, 
+                            zp=None, zpsys=None, shift=0.0,
+                            wave_original=None, trans_original=None):
     """Calculate bandflux for a single bandpass using precomputed static data.
     
     Parameters
@@ -563,18 +602,24 @@ def optimized_salt3_bandflux(phase, wave, dwave, trans, params, zp=None, zpsys=N
     dwave : float
         Spacing between wavelength grid points
     trans : array
-        Transmission values on the wavelength grid
+        Transmission values on the wavelength grid (used if shift=0)
     params : dict
         Dictionary containing model parameters: 'z', 't0', 'x0', 'x1', 'c'
     zp : float or None, optional
-        Zero point for flux scaling. If provided, zpsys must also be given
+        Zero point for flux scaling
     zpsys : str or None, optional
         Magnitude system (e.g. 'ab')
+    shift : float, optional
+        Constant wavelength shift to apply to transmission curve (in Angstroms)
+    wave_original : array, optional
+        Original wavelength array (required if shift != 0)
+    trans_original : array, optional
+        Original transmission array (required if shift != 0)
     
     Returns
     -------
     float or array
-        Flux in photons/s/cm^2. If phase is scalar then returns scalar
+        Flux in photons/s/cm^2
     """
     if zp is not None and zpsys is None:
         raise ValueError('zpsys must be given if zp is not None')
@@ -595,6 +640,23 @@ def optimized_salt3_bandflux(phase, wave, dwave, trans, params, zp=None, zpsys=N
 
     # Scale the integration grid to rest-frame wavelengths.
     restwave = wave * a
+    
+    # Get transmission values - use shifted version if shift is non-zero
[…]
