+++ examples/ns_multi_sn_shifts.py
@@ -1,0 +1,450 @@
+"""
+# Multi-Supernova Nested Sampling with Shared Transmission Shifts
+
+This script demonstrates how to fit multiple supernovae simultaneously with:
+- Individual SALT parameters (x0, x1, t0, c) for each supernova
+- Shared transmission shift parameters across all supernovae for each filter
+
+For more examples and the complete codebase, visit the [JAX-bandflux GitHub repository](https://github.com/samleeney/JAX-bandflux).
+"""
+
+import distrax
+import jax
+import jax.numpy as jnp
+import numpy as np
+import tqdm
+import blackjax
+import os
+from blackjax.ns.utils import log_weights
+from jax_supernovae.salt3 import optimized_salt3_multiband_flux
+from jax_supernovae.bandpasses import register_bandpass, get_bandpass
+from jax_supernovae.utils import save_chains_dead_birth
+from jax_supernovae.data import load_multiple_supernovae
+import matplotlib.pyplot as plt
+from anesthetic import read_chains, make_2d_axes
+
+# Configuration
+SN_NAMES = ['19dwz', '20aai']  # List of supernovae to fit
+fix_z = True  # Fix redshift from redshifts.dat or targets.dat
+fit_sigma = False  # Whether to fit an error scaling parameter
+
+# NS settings
+NS_SETTINGS = {
+    'n_delete': 1,
+    'n_live': 150,  # Increased for multi-SN fitting
+    'num_mcmc_steps_multiplier': 5
+}
+
+# Prior bounds
+PRIOR_BOUNDS = {
+    'z': {'min': 0.001, 'max': 0.2},
+    't0': {'min': 58000.0, 'max': 59000.0},
+    'x0': {'min': -5.0, 'max': -2.6},
+    'x1': {'min': -4.0, 'max': 4.0},
+    'c': {'min': -0.3, 'max': 0.3},
+    'log_sigma': {'min': -3.0, 'max': 1.0},
+    'shift': {'min': -50.0, 'max': 50.0}  # ±50 Å shift bounds
+}
+
+# Enable float64 precision
+jax.config.update("jax_enable_x64", True)
+
+# Load data for multiple supernovae
+print(f"Loading data for supernovae: {SN_NAMES}")
+multi_sn_data = load_multiple_supernovae(SN_NAMES, fix_z=fix_z)
+
+n_sne = multi_sn_data['n_sne']
+n_bands = multi_sn_data['n_bands']
+bridges = multi_sn_data['bridges']
+unique_bands = multi_sn_data['unique_bands']
+
+print(f"Loaded {n_sne} supernovae with {n_bands} unique bands: {unique_bands}")
+
+# Extract concatenated data arrays
+all_times = multi_sn_data['all_times']
+all_fluxes = multi_sn_data['all_fluxes']
+all_fluxerrs = multi_sn_data['all_fluxerrs']
+all_zps = multi_sn_data['all_zps']
+all_band_indices = multi_sn_data['all_band_indices']
+all_sn_indices = multi_sn_data['sn_indices']
+
+# Individual SN data for easier access
+times_list = multi_sn_data['times_list']
+fluxes_list = multi_sn_data['fluxes_list']
+fluxerrs_list = multi_sn_data['fluxerrs_list']
+zps_list = multi_sn_data['zps_list']
+band_indices_list = multi_sn_data['band_indices_list']
+fixed_z_list = multi_sn_data['fixed_z_list']
+
+# Define parameter structure
+# For N SNe: params[0:4N] are SN parameters (t0, log_x0, x1, c for each SN)
+# params[4N:4N+n_bands] are shared shift parameters
+# params[4N+n_bands] is optional log_sigma
+
+# Build parameter bounds and priors
[…]
