+++ jax_supernovae/data.py
@@ -109,8 +109,11 @@ def load_hsf_data(object_name, base_dir='data'):
     
     return data
 
-def load_redshift(object_name, redshift_file='data/redshifts.dat'):
-    """Load redshift for a given object from redshifts.dat.
+def load_redshift(object_name, redshift_file='data/redshifts.dat', targets_file='data/targets.dat'):
+    """Load redshift for a given object.
+    
+    First tries redshifts.dat (high-quality spectroscopic redshifts),
+    then falls back to targets.dat if object not found.
     
     Parameters
     ----------
@@ -118,64 +121,112 @@ def load_redshift(object_name, redshift_file='data/redshifts.dat'):
         Name of the object (e.g., '19agl')
     redshift_file : str
         Path to redshifts.dat file
+    targets_file : str
+        Path to targets.dat file (fallback)
         
     Returns
     -------
     tuple
         (redshift, redshift_err, flag) where:
         - redshift is the heliocentric redshift
-        - redshift_err is the symmetric error (max of plus/minus)
-        - flag is the reliability flag ('s'=strong, 'w'=weak, 'n'=no features)
+        - redshift_err is the symmetric error (max of plus/minus for redshifts.dat, 
+          or 0.001 default for targets.dat)
+        - flag is the reliability flag ('s'=strong, 'w'=weak, 'n'=no features,
+          or 'spu' from targets.dat)
             
     Raises
     ------
     FileNotFoundError
-        If redshift file not found
+        If neither redshift file nor targets file found
     ValueError
-        If object not found in redshift file
+        If object not found in either file
     """
-    # Try package data directory first
+    # First try redshifts.dat (primary source)
     package_redshift_file = os.path.join(PACKAGE_DIR, 'data', 'redshifts.dat')
     if os.path.exists(package_redshift_file):
         redshift_file = package_redshift_file
-    elif not os.path.exists(redshift_file):
-        raise FileNotFoundError(f"Redshift file not found: {redshift_file}")
         
-    # Skip comment lines and read data
-    with open(redshift_file, 'r') as f:
+    if os.path.exists(redshift_file):
+        # Skip comment lines and read data
+        with open(redshift_file, 'r') as f:
+            lines = f.readlines()
+        
+        data_lines = [l for l in lines if not l.startswith('#')]
+        
+        # Find all measurements for this object
+        measurements = []
+        for line in data_lines:
+            if not line.strip():
+                continue
+            parts = line.split()
+            if len(parts) < 6:
+                continue
+            if parts[0].lower() == object_name.lower():
+                try:
+                    z = float(parts[2])
+                    plus = float(parts[3])
+                    minus = float(parts[4])
+                    flag = parts[5] if len(parts) > 5 else 'n'
+                    measurements.append((z, plus, minus, flag))
+                except (ValueError, IndexError):
+                    continue
+        
+        if measurements:
+            # Prefer measurements with 's' flag, then 'w', then 'n'
+            flag_priority = {'s': 0, 'w': 1, 'n': 2}
+            measurements.sort(key=lambda x: flag_priority.get(x[3], 3))
+            
+            z, plus, minus, flag = measurements[0]
+            z_err = max(plus, minus)
+            
+            return z, z_err, flag
+    
+    # Fallback to targets.dat
+    package_targets_file = os.path.join(PACKAGE_DIR, 'data', 'targets.dat')
+    if os.path.exists(package_targets_file):
+        targets_file = package_targets_file
+    elif not os.path.exists(targets_file):
+        # Try in parent data directory
[…]
