+++ examples/ns_multi_sn.py
@@ -1,0 +1,334 @@
+"""
+Multi-supernova nested sampling for transmission shift calibration (Stage 1).
+
+This script implements Stage 1 of the nuisance-free supernova calibration,
+jointly fitting transmission shifts for filter bands across multiple SNe.
+"""
+
+import distrax
+import jax
+import jax.numpy as jnp
+import jax.lax as lax
+import numpy as np
+import tqdm
+import blackjax
+import os
+from blackjax.ns.utils import log_weights
+from jax_supernovae.salt3 import optimized_salt3_multiband_flux
+from jax_supernovae.utils import save_chains_dead_birth
+from jax_supernovae.multi_sn_utils import (
+    load_and_process_multiple_sne, 
+    unpack_parameters,
+    pack_parameters
+)
+import matplotlib.pyplot as plt
+from anesthetic import read_chains, make_2d_axes
+
+# Enable float64 precision
+jax.config.update("jax_enable_x64", True)
+
+# Configuration
+SN_NAMES = ['19agl', '19ahi', '19ahr']  # Start with 3 SNe for testing
+FIX_Z = True  # Fix redshift for all SNe
+
+# Nested sampling settings
+NS_SETTINGS = {
+    'n_delete': 1,
+    'n_live': 200,  # Increased for more parameters
+    'num_mcmc_steps_multiplier': 5
+}
+
+# Prior bounds for transmission shifts (in Angstroms)
+T_SHIFT_BOUNDS = {'min': -50.0, 'max': 50.0}
+
+# Prior bounds for SN parameters (same as single SN case)
+SN_PARAM_BOUNDS = {
+    't0': {'min': 58000.0, 'max': 59000.0},
+    'x0': {'min': -5.0, 'max': -2.6},
+    'x1': {'min': -4.0, 'max': 4.0},
+    'c': {'min': -0.3, 'max': 0.3},
+}
+
+print(f"Loading data for {len(SN_NAMES)} supernovae...")
+sne_data, global_band_names, global_bridges = load_and_process_multiple_sne(
+    SN_NAMES, data_dir='data', fix_z=FIX_Z
+)
+
+# Define dimensions
+n_bands_global = len(global_band_names)
+n_sne = len(SN_NAMES)
+n_params_sn = 4  # t0, log_x0, x1, c per SN
+n_params_total = n_bands_global + n_sne * n_params_sn
+
+print(f"Global bands: {global_band_names}")
+print(f"Total parameters: {n_params_total}")
+print(f"  - Transmission shifts: {n_bands_global}")
+print(f"  - SN parameters: {n_sne * n_params_sn} ({n_sne} SNe x {n_params_sn} params)")
+
+# Create prior distributions
+prior_dists = []
+
+# Transmission shift priors (uniform)
+for band_name in global_band_names:
+    prior_dists.append(
+        distrax.Uniform(low=T_SHIFT_BOUNDS['min'], high=T_SHIFT_BOUNDS['max'])
+    )
+
+# SN parameter priors (uniform for each SN)
+for i in range(n_sne):
+    prior_dists.append(distrax.Uniform(low=SN_PARAM_BOUNDS['t0']['min'], 
+                                      high=SN_PARAM_BOUNDS['t0']['max']))
+    prior_dists.append(distrax.Uniform(low=SN_PARAM_BOUNDS['x0']['min'], 
+                                      high=SN_PARAM_BOUNDS['x0']['max']))
+    prior_dists.append(distrax.Uniform(low=SN_PARAM_BOUNDS['x1']['min'], 
+                                      high=SN_PARAM_BOUNDS['x1']['max']))
+    prior_dists.append(distrax.Uniform(low=SN_PARAM_BOUNDS['c']['min'], 
[…]
