+++ jax_supernovae/multi_sn_utils.py
@@ -1,0 +1,208 @@
+"""Utility functions for multi-supernova nested sampling with transmission calibration."""
+import jax
+import jax.numpy as jnp
+import numpy as np
+from jax_supernovae.data import load_and_process_data, load_hsf_data
+from jax_supernovae.bandpasses import register_all_bandpasses
+
+jax.config.update("jax_enable_x64", True)
+
+def load_and_process_multiple_sne(sn_names, data_dir='data', fix_z=True):
+    """Loads and processes data for multiple SNe, preparing it for JIT compilation.
+    
+    Parameters
+    ----------
+    sn_names : list of str
+        List of supernova names to load
+    data_dir : str
+        Directory containing the data files
+    fix_z : bool
+        Whether to fix redshift
+        
+    Returns
+    -------
+    tuple
+        (sne_data_jax, global_band_names, global_bridges) where:
+        - sne_data_jax: PyTree of padded JAX arrays containing all data
+        - global_band_names: List of unique band names across all SNe
+        - global_bridges: Tuple of bridge data for each global band
+    """
+    # Register all standard bandpasses and get their bridges
+    _, bridges_dict = register_all_bandpasses()
+    
+    # 1. First pass: Collect all data and determine global properties
+    all_sn_data = []
+    global_band_names = []
+    max_n_points = 0
+    max_n_bands_per_sn = 0
+    all_sn_bands = []  # Store band names for each SN
+
+    for sn_name in sn_names:
+        # Load processed data
+        data = load_and_process_data(sn_name, data_dir, fix_z)
+        all_sn_data.append(data)
+        
+        # Load raw data to get band names
+        phot_data = load_hsf_data(sn_name, base_dir=data_dir)
+        sn_unique_bands = np.unique(phot_data['band']).tolist()
+        all_sn_bands.append(sn_unique_bands)
+        
+        # Add to global band list
+        for band in sn_unique_bands:
+            if band not in global_band_names:
+                global_band_names.append(band)
+        
+        # Track maximums for padding
+        max_n_points = max(max_n_points, len(data[0]))
+        max_n_bands_per_sn = max(max_n_bands_per_sn, len(sn_unique_bands))
+    
+    # Sort global band names for consistency
+    global_band_names = sorted(global_band_names)
+    global_band_map = {name: i for i, name in enumerate(global_band_names)}
+    
+    # Create tuple of bridges in global band order
+    global_bridges = []
+    for name in global_band_names:
+        if name in bridges_dict:
+            global_bridges.append(bridges_dict[name])
+        else:
+            print(f"Warning: Bridge not found for band {name}")
+    global_bridges = tuple(global_bridges)
+    
+    n_sne = len(sn_names)
+    
+    # 2. Second pass: Create padded arrays
+    sne_data = {
+        'times': np.full((n_sne, max_n_points), -1.0),
+        'fluxes': np.zeros((n_sne, max_n_points)),
+        'fluxerrs': np.full((n_sne, max_n_points), 1e10),  # Large error for invalid points
+        'zps': np.zeros((n_sne, max_n_points)),
+        'band_indices': np.full((n_sne, max_n_points), -1, dtype=int),
+        'valid_mask': np.zeros((n_sne, max_n_points), dtype=bool),
+        'local_to_global_map': np.full((n_sne, max_n_bands_per_sn), -1, dtype=int),
+        'fixed_z': np.zeros(n_sne),
+        'n_points': np.zeros(n_sne, dtype=int),
+        'n_bands': np.zeros(n_sne, dtype=int),
+        'sn_bridges': []  # Will store per-SN bridge tuples
+    }
+    
+    for i, sn_name in enumerate(sn_names):
+        times, fluxes, fluxerrs, zps, band_indices, bridges, fixed_z = all_sn_data[i]
[…]

[…]
