### 2025-10-10 2c12cf2
Fix CI: Update Python version to 3.11+ for Distrax compatibility
- Update GitHub Actions workflows to use Python 3.11
- Update setup.py to require Python >=3.11
- Required for Distrax dependency which now requires Python 3.11+
