+++ parameter_statistics.txt
@@ -1,9 +1,0 @@
-Parameter Statistics:
---------------------------------------------------
-Parameter               Mean         Std Dev
---------------------------------------------------
-t0              58606.506972        0.609268
-log_x0             -2.967710        0.022792
-x1                  1.400588        0.713620
-c                   0.055117        0.064023
---------------------------------------------------
