### 2025-10-10 581e035
Add v3.0 functional API tests and update existing tests
- Add comprehensive test suite for v3.0 functional API (test_api_compatibility.py)
  - Tests parameter handling, broadcasting semantics, return types
  - Tests magnitude calculations and property methods
  - 24 tests covering all API compatibility requirements

- Update existing tests for v3.0 functional API
  - test_bandflux_performance.py: Update to use params dict
  - test_documentation.py: Update examples to v3.0 API
  - test_dust_comparison.py: Update to functional parameters
  - test_dust_extinction.py: Update to functional paramete
[…]
