+++ tests/test_dust_extinction.py
@@ -1,4 +1,17 @@
-"""Tests for dust extinction laws in JAX-bandflux."""
+"""Tests for dust extinction laws in JAX-bandflux.
+
+NOTE: The SALT3Source class does not yet expose dust parameters directly.
+For now, dust extinction tests use the low-level optimized_salt3_bandflux()
+function with dust parameters. In a future version, dust support will be
+added to the Source-level API.
+
+TODO: Once dust parameters are exposed in SALT3Source, update tests to use:
+    source = SALT3Source()
+    source['x0'] = 1e-5
+    source['ebv'] = 0.1  # Future feature
+    source['r_v'] = 3.1  # Future feature
+    source['dust_type'] = 'CCM89'  # Future feature
+"""
 
 import numpy as np
 import jax
@@ -17,38 +30,38 @@ def test_dust_extinction_laws():
     """Test that JAX dust extinction laws match sncosmo's implementation."""
     # Test wavelength grid
     wave = np.linspace(3000, 10000, 100)
-    
+
     # Test parameters
     ebv = 0.1
     r_v = 3.1
-    
+
     # Test CCM89 dust law
     ccm89_jax = dust.ccm89_extinction(wave, ebv, r_v)
     ccm89_sncosmo = sncosmo.CCM89Dust()
     ccm89_sncosmo.parameters = [ebv, r_v]
     ccm89_sncosmo_flux = ccm89_sncosmo.propagate(wave, np.ones_like(wave))
     ccm89_sncosmo_extinction = -2.5 * np.log10(ccm89_sncosmo_flux)
-    
+
     # Check that the results are close
     np.testing.assert_allclose(ccm89_jax, ccm89_sncosmo_extinction, rtol=1e-3)
-    
+
     # Test OD94 dust law
     od94_jax = dust.od94_extinction(wave, ebv, r_v)
     od94_sncosmo = sncosmo.OD94Dust()
     od94_sncosmo.parameters = [ebv, r_v]
     od94_sncosmo_flux = od94_sncosmo.propagate(wave, np.ones_like(wave))
     od94_sncosmo_extinction = -2.5 * np.log10(od94_sncosmo_flux)
-    
+
     # Check that the results are close
     np.testing.assert_allclose(od94_jax, od94_sncosmo_extinction, rtol=1e-3)
-    
+
     # Test F99 dust law
     f99_jax = dust.f99_extinction(wave, ebv, r_v)
     f99_sncosmo = sncosmo.F99Dust(r_v=r_v)
     f99_sncosmo.parameters = [ebv]
     f99_sncosmo_flux = f99_sncosmo.propagate(wave, np.ones_like(wave))
     f99_sncosmo_extinction = -2.5 * np.log10(f99_sncosmo_flux)
-    
+
     # Check that the results are close with a higher tolerance for F99
     # F99 is more complex and our implementation is based on interpolation
     np.testing.assert_allclose(f99_jax, f99_sncosmo_extinction, rtol=6e-2)
@@ -58,37 +71,37 @@ def test_dust_extinction_laws_uv():
     """Test dust extinction laws in the UV range where issues were occurring."""
     # Test wavelength grid focusing on UV range
     wave = np.linspace(2000, 3500, 100)
-    
+
     # Test parameters
     ebv = 0.1
     r_v = 3.1
-    
+
     # Test CCM89 dust law
     ccm89_jax = dust.ccm89_extinction(wave, ebv, r_v)
     ccm89_sncosmo = sncosmo.CCM89Dust()
     ccm89_sncosmo.parameters = [ebv, r_v]
[…]
