+++ tests/test_transmission_shifts.py
@@ -1,103 +1,311 @@
+"""Test transmission shift functionality for bandpasses.
+
+This module tests wavelength-dependent transmission shifts in bandpass filters.
+These shifts can be used to model systematic effects or instrumental variations.
+
+NOTE: Transmission shifts are not yet exposed in the SALT3Source API.
+These tests verify the underlying optimized functions still work correctly.
+TODO: Add transmission shift support to SALT3Source.bandflux() in future release.
+"""
+
 import pytest
+import numpy as np
 import jax.numpy as jnp
 from jax_supernovae.bandpasses import Bandpass
 from jax_supernovae.salt3 import optimized_salt3_bandflux, precompute_bandflux_bridge
 
-def test_constant_shift():
-    """Test constant wavelength shift."""
-    # Create test bandpass with finer sampling for accurate peak detection
+
+def test_bandpass_constant_shift():
+    """Test that constant wavelength shift works at bandpass level.
+
+    A constant shift should move the transmission peak by approximately
+    the shift amount. This is the fundamental test for shift functionality.
+    """
+    # Create test bandpass with Gaussian transmission
     wave = jnp.linspace(4000, 7000, 1000)
     trans = jnp.exp(-(wave - 5500)**2 / (2 * 500**2))
     bp = Bandpass(wave, trans)
-    
-    # Test shifted transmission
-    shift = 50.0  # 50 Å shift
+
+    # Apply shift
+    shift = 50.0  # 50 Angstrom shift
     trans_original = bp(wave)
     trans_shifted = bp(wave, shift=shift)
-    
-    # When we apply a positive shift, the effective wavelength is wave - shift,
-    # so the transmission peak appears to move to the right
+
+    # Find peaks
     peak_original = wave[jnp.argmax(trans_original)]
     peak_shifted = wave[jnp.argmax(trans_shifted)]
-    
-    # The peak should move by approximately +shift
-    # Allow for some discretization error
-    assert jnp.abs(peak_shifted - peak_original - shift) < 5.0
 
-def test_flux_calculation_with_shift():
-    """Test flux calculation with transmission shift."""
-    # Create narrower Gaussian bandpass for more sensitivity to shifts
+    # Peak should move by approximately +shift
+    # (Positive shift means effective wavelength is wave - shift,
+    # so transmission peak appears to move right)
+    peak_diff = peak_shifted - peak_original
+    assert jnp.abs(peak_diff - shift) < 5.0, \
+        f"Peak moved by {peak_diff:.1f} Å, expected ~{shift:.1f} Å"
+
+    print(f"\n✓ Constant shift test passed:")
+    print(f"  Shift applied: {shift:.1f} Å")
+    print(f"  Peak moved: {peak_diff:.1f} Å")
+
+
+def test_flux_with_transmission_shift():
+    """Test flux calculation with transmission shift.
+
+    This tests the underlying optimized_salt3_bandflux function with shift
+    parameters. The shifted and unshifted fluxes should be detectably different.
+    """
+    # Create narrower Gaussian bandpass for sensitivity to shifts
     wave = jnp.linspace(4000, 7000, 300)
-    trans = jnp.exp(-(wave - 5500)**2 / (2 * 200**2))  # Narrower bandpass
+    trans = jnp.exp(-(wave - 5500)**2 / (2 * 200**2))
     bp = Bandpass(wave, trans)
-    
+
     # Prepare bridge with shift support
     bridge = precompute_bandflux_bridge(bp)
-    
-    # Test parameters
+
+    # Model parameters
[…]
