### 2025-10-10 c2c96d2
Add SALT3Source class with v3.0 functional API
- Create jax_supernovae/source.py with SALT3Source class
- Implement bandflux() method with functional parameter passing
- Update __init__.py to export SALT3Source
- Fixes ImportError when importing SALT3Source from jax_supernovae
