+++ jax_supernovae/__init__.py
@@ -1,1 +1,3 @@
- 
+from jax_supernovae.source import SALT3Source
+
+__all__ = ['SALT3Source']
