### 2025-10-10 a460552
Update SALT3Source and examples to use v3.0 functional API
- Modify load_and_process_data() to return unique_bands list
- Redesign SALT3Source.bandflux() to accept precomputed bridges
- Update examples/ns.py to use SALT3Source with functional API
- Calculate phases from times in likelihood function
- Pass params as dict to source.bandflux() (v3.0 functional API)
