+++ jax_supernovae/source.py
@@ -7,8 +7,9 @@ the source object.
 """
 
 import jax.numpy as jnp
+import numpy as np
 from jax_supernovae.salt3 import optimized_salt3_multiband_flux
-from jax_supernovae.bandpasses import get_bandpass, register_all_bandpasses
+from jax_supernovae.bandpasses import register_all_bandpasses
 
 
 class SALT3Source:
@@ -24,11 +25,24 @@ class SALT3Source:
 
     Examples
     --------
+    Basic usage with string band names:
     >>> source = SALT3Source(name='salt3-nir')
     >>> params = {'x0': 1e-5, 'x1': 0.0, 'c': 0.0}
-    >>> bands = ['g', 'r', 'i', 'z']
-    >>> phases = jnp.array([-10.0, 0.0, 10.0, 20.0])
-    >>> fluxes = source.bandflux(params, bands, phases, zp=27.5, zpsys='ab')
+    >>> flux = source.bandflux(params, 'bessellb', 0.0, zp=27.5, zpsys='ab')
+
+    High-performance usage with precomputed data (for nested sampling):
+    >>> from jax_supernovae.data import load_and_process_data
+    >>> times, fluxes, fluxerrs, zps, band_indices, bands, bridges, fixed_z = \\
+    ...     load_and_process_data('19dwz', fix_z=True)
+    >>> source = SALT3Source()
+    >>> # Inside likelihood function:
+    >>> t0 = 58650.0
+    >>> z = fixed_z[0]
+    >>> phases = (times - t0) / (1 + z)
+    >>> band_names = [bands[i] for i in band_indices]
+    >>> fluxes = source.bandflux(params, band_names, phases, zp=zps, zpsys='ab',
+    ...                          band_indices=band_indices, bridges=bridges,
+    ...                          unique_bands=bands)
     """
 
     def __init__(self, name='salt3-nir'):
@@ -46,7 +60,8 @@ class SALT3Source:
         # Register all bandpasses to ensure they're available
         register_all_bandpasses()
 
-    def bandflux(self, params, bands, phases, zp=None, zpsys=None):
+    def bandflux(self, params, bands, phases, zp=None, zpsys=None,
+                 band_indices=None, bridges=None, unique_bands=None):
         """Calculate bandflux using v3.0 functional API.
 
         Parameters
@@ -62,28 +77,35 @@ class SALT3Source:
         bands : str or array-like
             Bandpass name(s). Can be a string for a single band or array of band names.
         phases : float or array
-            Rest-frame phase(s) relative to t0
+            Rest-frame phase(s) relative to t0.
         zp : float or array, optional
             Zero point(s) for flux scaling
         zpsys : str, optional
             Zero point system ('ab' or 'vega')
+        band_indices : array, optional
+            For performance: indices into unique_bands/bridges arrays.
+            If provided, must also provide bridges and unique_bands.
+        bridges : tuple, optional
+            For performance: precomputed bridge data structures.
+        unique_bands : list, optional
+            For performance: list of unique band names corresponding to bridges.
 
         Returns
         -------
         flux : array
-            Bandflux value(s) with shape matching the broadcast shape of inputs
+            Bandflux value(s) with shape matching the input
 
         Examples
         --------
         Single band and phase:
         >>> source = SALT3Source()
         >>> params = {'x0': 1e-5, 'x1': 0.0, 'c': 0.0}
-        >>> flux = source.bandflux(params, 'g', 0.0, zp=27.5, zpsys='ab')
+        >>> flux = source.bandflux(params, 'bessellb', 0.0, zp=27.5, zpsys='ab')
 
-        Multiple bands and phases:
[…]
