+++ parameter_statistics.txt
@@ -1,0 +1,9 @@
+Parameter Statistics:
+--------------------------------------------------
+Parameter               Mean         Std Dev
+--------------------------------------------------
+t0              58606.512942        0.654155
+log_x0             -2.966406        0.025186
+x1                  1.433097        0.711414
+c                   0.048316        0.070168
+--------------------------------------------------
