### 2025-10-10 497094e
Add documentation explaining API differences from SNCosmo
- Add concise API compatibility note to README
- Create detailed docs/api_differences.rst explaining:
  a) Functional parameter API (no dict-style access)
  b) Optional bridges parameter for performance
- Include migration guide and performance comparison
- Add to docs toctree and quick links
