+++ docs/api_differences.rst
@@ -1,0 +1,296 @@
+API Differences from SNCosmo
+============================
+
+JAX-bandflux provides a SALT3Source implementation that aims to be as similar as possible to SNCosmo's API, while accommodating JAX's requirements for just-in-time (JIT) compilation and GPU acceleration.
+
+Design Philosophy
+-----------------
+
+We've designed JAX-bandflux to:
+
+1. **Maintain numerical consistency**: Results match SNCosmo within 0.001%
+2. **Enable JIT compilation**: All code paths are compatible with ``@jax.jit``
+3. **Maximize performance**: Precomputation strategies for repeated calculations
+4. **Minimize API changes**: Keep the interface as familiar as possible
+
+Key Differences
+---------------
+
+1. Functional Parameter API
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+**SNCosmo approach** (stateful):
+
+.. code-block:: python
+
+   import sncosmo
+
+   # Parameters stored in the source object
+   source = sncosmo.get_source('salt3-nir')
+   source['z'] = 0.1
+   source['t0'] = 58650.0
+   source['x0'] = 1e-5
+   source['x1'] = 0.0
+   source['c'] = 0.0
+
+   # Bandflux uses stored parameters
+   flux = source.bandflux('bessellb', 0.0, zp=27.5, zpsys='ab')
+
+**JAX-bandflux approach** (functional):
+
+.. code-block:: python
+
+   from jax_supernovae import SALT3Source
+
+   # Parameters passed as dictionary argument
+   source = SALT3Source()
+   params = {
+       'x0': 1e-5,
+       'x1': 0.0,
+       'c': 0.0
+   }
+
+   # Bandflux receives parameters as argument
+   flux = source.bandflux(params, 'bessellb', 0.0, zp=27.5, zpsys='ab')
+
+**Why this difference?**
+
+JAX's JIT compiler cannot handle mutable object state. By passing parameters as function arguments rather than storing them as attributes, we enable:
+
+- JIT compilation of the entire likelihood function
+- Automatic differentiation through the model
+- GPU acceleration without state management issues
+
+The functional API is a requirement for JAX compatibility, not a design choice.
+
+2. Precomputed Bridges for Performance
+~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+
+**SNCosmo approach** (bandpass by name):
+
+.. code-block:: python
+
+   import sncosmo
+
+   source = sncosmo.get_source('salt3-nir')
+   source.set(z=0.1, t0=58650.0, x0=1e-5, x1=0.0, c=0.0)
+
+   # Bandpass loaded/interpolated each time
+   for i in range(100000):  # Nested sampling
+       flux = source.bandflux('bessellb', phases[i], zp=27.5, zpsys='ab')
+
+**JAX-bandflux approach** (with precomputed bridges):
+
+.. code-block:: python
+
+   from jax_supernovae import SALT3Source
+   from jax_supernovae.data import load_and_process_data
+
+   # Load data and precompute bridges ONCE
+   times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = \\
+       load_and_process_data('19dwz', fix_z=True)
+
+   source = SALT3Source()
+   params = {'x0': 1e-5, 'x1': 0.0, 'c': 0.0}
+
+   # Calculate all fluxes using precomputed bridges
+   phases = (times - t0) / (1 + z)
+   fluxes = source.bandflux(
+       params, None, phases, zp=zps, zpsys='ab',
+       band_indices=band_indices,    # Which filter for each observation
[…]
