### 2025-10-10 15c55f4
Complete SALT3Source v3.0 API implementation with full compatibility
- Fix param_names to include all 5 SALT3 parameters (z, t0, x0, x1, c)
- Implement standard bandflux path for simple calculations without bridges
- Add bandmag() method for magnitude calculations
- Add properties: minphase, maxphase, minwave, maxwave, __str__, __repr__
- Clean up examples: remove outdated ns_original.py
- Update .gitignore: add docs/_build/ and parameter_statistics.txt

The SALT3Source class now provides:
1. High-performance path with precomputed bridges (for nested sampling)
2. Standard path with on-the-fly 
[…]
