### 2025-10-10 01cd20c
Simplify ns.py example - reduce from 384 to 167 lines
Simplifications:
- Assume fixed z (remove all fix_z conditionals)
- Simplify prior setup using dictionary comprehensions
- Reduce logprior/loglikelihood to essential logic
- Simplify sample_from_priors with list comprehension
- Remove verbose print statements
- Streamline results processing and plotting

The example now focuses on the core nested sampling workflow
without excessive branching logic.
