+++ examples/ns_anomaly.py
@@ -1,149 +1,85 @@
 """
 Anomaly detection for supernova light curves using nested sampling.
 
-This script implements a Bayesian anomaly detection framework for supernova
-light curve analysis. It runs two nested sampling procedures:
-1. A standard version that fits SALT3 model parameters to the data
-2. An anomaly detection version that includes an additional parameter (log_p)
-   to identify potential outliers in the data
-
-The script generates comparison plots, parameter statistics, and weighted
-anomaly indicators (emax values) that highlight potential problematic data points.
+Runs two nested sampling procedures:
+1. Standard: fits SALT3 model parameters
+2. Anomaly: includes log_p parameter to identify outliers
 """
-import distrax
 import os
 os.environ['XLA_PYTHON_CLIENT_PREALLOCATE'] = 'false'
+import distrax
 import jax
 import jax.numpy as jnp
 import numpy as np
 import tqdm
 import blackjax
 import yaml
-import pandas as pd
 from blackjax.ns.utils import log_weights
 from jax_supernovae.salt3 import optimized_salt3_multiband_flux, precompute_bandflux_bridge
-from jax_supernovae.bandpasses import register_bandpass, get_bandpass, register_all_bandpasses, Bandpass
-from jax_supernovae.utils import save_chains_dead_birth
+from jax_supernovae.bandpasses import register_bandpass, register_all_bandpasses, Bandpass
 from jax_supernovae.data import load_redshift, load_hsf_data
 import matplotlib.pyplot as plt
 from anesthetic import read_chains, make_2d_axes
-import requests
 
-
-# Define default settings for nested sampling and prior bounds
-DEFAULT_NS_SETTINGS = {
-    'max_iterations': int(os.environ.get('NS_MAX_ITERATIONS', '500')),
+# Configuration
+DEFAULT_SETTINGS = {
+    'sn_name': '19vnk',
+    'selected_bandpasses': None,
+    'custom_bandpass_files': None,
+    'max_iterations': 500,
     'n_delete': 75,
     'n_live': 150,
     'num_mcmc_steps_multiplier': 5,
-    'fit_sigma': False,
-    'fit_log_p': True,
-    'fit_z': True
+    'fit_sigma': False
 }
 
 DEFAULT_PRIOR_BOUNDS = {
-    'z': {'min': 0.001, 'max': 0.2},
-    't0': {'min': 58000.0, 'max': 60000.0},
-    'x0': {'min': -5.0, 'max': -1},
-    'x1': {'min': -10, 'max': 10},
-    'c': {'min': -0.6, 'max': 0.6},
-    'sigma': {'min': 0.001, 'max': 5},
-    'log_p': {'min': -20, 'max': -1}
+    't0': (58000.0, 60000.0),
+    'x0': (-5.0, -1.0),
+    'x1': (-10.0, 10.0),
+    'c': (-0.6, 0.6),
+    'sigma': (0.001, 5.0),
+    'log_p': (-20.0, -1.0)
 }
 
-# Default settings
-DEFAULT_SETTINGS = {
-    'fix_z': True,
-    'sn_name': '19vnk',  # Default supernova to analyze
-    'selected_bandpasses': None,  # Default: use all available bandpasses
-    'custom_bandpass_files': None  # Default: no custom bandpass files
-}
-
-# Try to load settings.yaml; if not found, use an empty dictionary
+# Load settings from file if available
 try:
     with open('settings.yaml', 'r') as f:
-        settings_from_file = yaml.safe_load(f)
+        settings_from_file = yaml.safe_load(f) or {}
 except FileNotFoundError:
     settings_from_file = {}
 
-# Merge the settings from file with the defaults
[…]
