+++ examples/ns.py
@@ -1,11 +1,9 @@
 """
 # Nested Sampling with JAX-bandflux
 
-This script demonstrates how to run the nested sampling procedure for supernovae SALT model fitting using the JAX-bandflux package. We will install the package, load the data, set up and run the nested sampling algorithm, and finally produce a corner plot of the posterior samples.
-
-For more examples and the complete codebase, visit the [JAX-bandflux GitHub repository](https://github.com/samleeney/JAX-bandflux). The academic paper associated with this work can be found [here](https://github.com/samleeney/JAX-bandflux/blob/71ca8d1b3b273147e1e9bf60a9ef11a806363b80/paper.bib).
+This script demonstrates nested sampling for SALT3 supernova model fitting.
+The redshift is fixed (not fitted) and flux uncertainties can optionally be scaled.
 """
-
 
 import distrax
 import jax
@@ -13,7 +11,6 @@ import jax.numpy as jnp
 import numpy as np
 import tqdm
 import blackjax
-import os
 from blackjax.ns.utils import log_weights
 from jax_supernovae import SALT3Source
 from jax_supernovae.utils import save_chains_dead_birth
@@ -21,14 +18,8 @@ from jax_supernovae.data import load_and_process_data
 import matplotlib.pyplot as plt
 from anesthetic import read_chains, make_2d_axes
 
-# Option to fit sigma or not.
-# When fit_sigma is True, the code will include an extra free parameter (log_sigma)
-# so that the effective flux error becomes sigma * flux_err with sigma = 10**(log_sigma).
-# When fit_sigma is False, the code will use the measured flux_err from the file.
-fit_sigma = False
-
-# Settings that were previously in YAML
-fix_z = True
+# Configuration
+fit_sigma = False  # Fit an extra parameter to scale flux uncertainties
 
 NS_SETTINGS = {
     'n_delete': 60,
@@ -36,237 +27,76 @@ NS_SETTINGS = {
     'num_mcmc_steps_multiplier': 5
 }
 
-# Prior bounds (added log_sigma bound)
 PRIOR_BOUNDS = {
-    'z': {'min': 0.001, 'max': 0.2},
-    't0': {'min': 58000.0, 'max': 59000.0},
-    'x0': {'min': -5.0, 'max': -2.6},
-    'x1': {'min': -4.0, 'max': 4.0},
-    'c': {'min': -0.3, 'max': 0.3},
-    'log_sigma': {'min': -3.0, 'max': 1.0}
+    't0': (58000.0, 59000.0),
+    'x0': (-5.0, -2.6),
+    'x1': (-4.0, 4.0),
+    'c': (-0.3, 0.3),
+    'log_sigma': (-3.0, 1.0)
 }
 
 # Enable float64 precision
 jax.config.update("jax_enable_x64", True)
 
-# Load and process data
-times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = load_and_process_data('19dwz', data_dir='data', fix_z=fix_z)
+# Load data with fixed redshift
+times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = \
+    load_and_process_data('19dwz', data_dir='data', fix_z=True)
 
-# Create SALT3 source for bandflux calculations
+z = fixed_z[0]
 source = SALT3Source()
 
-# Define parameter bounds and priors
-if fix_z:
-    param_bounds = {
-        't0': (PRIOR_BOUNDS['t0']['min'], PRIOR_BOUNDS['t0']['max']),
[…]
