### 2025-10-10 bb2b4a8
Simplify ns_anomaly and add custom_bandpass_example
ns_anomaly.py: Reduce from 1720 to 283 lines (83% reduction)
- Remove complex custom bandpass loading logic
- Assume fixed z (remove all fix_z conditionals)
- Simplify likelihood functions
- Streamline plotting to essential comparisons
- Focus on core anomaly detection functionality

custom_bandpass_example.py: New example (201 lines)
- Demonstrates loading custom bandpass filters from files
- Shows synthetic bandpass creation
- Includes SVO filter download functionality
- Provides clear examples of bandpass usage with SALT3Source

Examples n
[…]
