+++ examples/ns_anomaly.py
@@ -1,9 +1,12 @@
 """
-Anomaly detection for supernova light curves using nested sampling.
+Anomaly Detection for Supernova Light Curves
 
 Runs two nested sampling procedures:
 1. Standard: fits SALT3 model parameters
 2. Anomaly: includes log_p parameter to identify outliers
+
+The anomaly model assigns each data point a probability of being an outlier,
+allowing identification of problematic observations.
 """
 import os
 os.environ['XLA_PYTHON_CLIENT_PREALLOCATE'] = 'false'
@@ -13,1644 +16,268 @@ import jax.numpy as jnp
 import numpy as np
 import tqdm
 import blackjax
-import yaml
 from blackjax.ns.utils import log_weights
-from jax_supernovae.salt3 import optimized_salt3_multiband_flux, precompute_bandflux_bridge
-from jax_supernovae.bandpasses import register_bandpass, register_all_bandpasses, Bandpass
-from jax_supernovae.data import load_redshift, load_hsf_data
+from jax_supernovae.salt3 import optimized_salt3_multiband_flux
+from jax_supernovae.data import load_and_process_data
 import matplotlib.pyplot as plt
 from anesthetic import read_chains, make_2d_axes
 
 # Configuration
-DEFAULT_SETTINGS = {
-    'sn_name': '19vnk',
-    'selected_bandpasses': None,
-    'custom_bandpass_files': None,
-    'max_iterations': 500,
-    'n_delete': 75,
-    'n_live': 150,
-    'num_mcmc_steps_multiplier': 5,
-    'fit_sigma': False
-}
-
-DEFAULT_PRIOR_BOUNDS = {
+PRIOR_BOUNDS = {
     't0': (58000.0, 60000.0),
     'x0': (-5.0, -1.0),
     'x1': (-10.0, 10.0),
     'c': (-0.6, 0.6),
-    'sigma': (0.001, 5.0),
     'log_p': (-20.0, -1.0)
 }
 
-# Load settings from file if available
-try:
-    with open('settings.yaml', 'r') as f:
-        settings_from_file = yaml.safe_load(f) or {}
-except FileNotFoundError:
-    settings_from_file = {}
+NS_SETTINGS = {
+    'n_delete': 75,
+    'n_live': 150,
+    'num_mcmc_steps_multiplier': 5,
+    'max_iterations': 500
+}
 
-settings = {**DEFAULT_SETTINGS, **settings_from_file}
-PRIOR_BOUNDS = {**DEFAULT_PRIOR_BOUNDS, **settings_from_file.get('prior_bounds', {})}
-
-sn_name = settings['sn_name']
-selected_bandpasses = settings['selected_bandpasses']
-custom_bandpass_files = settings['custom_bandpass_files']
-fit_sigma = settings['fit_sigma']
+SN_NAME = '19dwz'
+DATA_DIR = 'data'
 
 jax.config.update("jax_enable_x64", True)
 
-def create_wfcam_j_bandpass():
-    """Load WFCAM J filter bandpass from file."""
-    filter_paths = [
-        os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))),
-                     'filter_data', 'UKIRT_WFCAM.J.dat'),
-        'WFCAM_J.dat',
-        'UKIRT_WFCAM.J.dat',
-        os.path.join(os.path.dirname(os.path.abspath(__file__)), 'WFCAM_J.dat'),
-        os.path.join(os.path.dirname(os.path.abspath(__file__)), 'UKIRT_WFCAM.J.dat'),
-        os.path.join(os.path.dirname(os.path.abspath(__file__)), 'filter_data/UKIRT_WFCAM.J.dat'),
-    ]
+# Load data with fixed redshift
[…]
