### 2025-10-10 bae0ec9
Fix shape issues in prior/likelihood functions
- Use jnp.stack and jnp.sum instead of Python sum() for proper JAX array handling
- Remove unnecessary reshape operations that were adding extra dimensions
- Ensures logprior and loglikelihood return 1D arrays as expected by BlackJAX
- Fixes ValueError: k argument to top_k error in nested sampling
