### 2025-10-10 e96089c
Revert examples to working versions before simplification
Restored ns.py and ns_anomaly.py to their state at commit 15c55f4
where they were known to work correctly. The simplification changes
introduced shape issues that broke the nested sampling.
