+++ examples/ns.py
@@ -1,9 +1,11 @@
 """
 # Nested Sampling with JAX-bandflux
 
-This script demonstrates nested sampling for SALT3 supernova model fitting.
-The redshift is fixed (not fitted) and flux uncertainties can optionally be scaled.
+This script demonstrates how to run the nested sampling procedure for supernovae SALT model fitting using the JAX-bandflux package. We will install the package, load the data, set up and run the nested sampling algorithm, and finally produce a corner plot of the posterior samples.
+
+For more examples and the complete codebase, visit the [JAX-bandflux GitHub repository](https://github.com/samleeney/JAX-bandflux). The academic paper associated with this work can be found [here](https://github.com/samleeney/JAX-bandflux/blob/71ca8d1b3b273147e1e9bf60a9ef11a806363b80/paper.bib).
 """
+
 
 import distrax
 import jax
@@ -11,6 +13,7 @@ import jax.numpy as jnp
 import numpy as np
 import tqdm
 import blackjax
+import os
 from blackjax.ns.utils import log_weights
 from jax_supernovae import SALT3Source
 from jax_supernovae.utils import save_chains_dead_birth
@@ -18,8 +21,14 @@ from jax_supernovae.data import load_and_process_data
 import matplotlib.pyplot as plt
 from anesthetic import read_chains, make_2d_axes
 
-# Configuration
-fit_sigma = False  # Fit an extra parameter to scale flux uncertainties
+# Option to fit sigma or not.
+# When fit_sigma is True, the code will include an extra free parameter (log_sigma)
+# so that the effective flux error becomes sigma * flux_err with sigma = 10**(log_sigma).
+# When fit_sigma is False, the code will use the measured flux_err from the file.
+fit_sigma = False
+
+# Settings that were previously in YAML
+fix_z = True
 
 NS_SETTINGS = {
     'n_delete': 60,
@@ -27,77 +36,237 @@ NS_SETTINGS = {
     'num_mcmc_steps_multiplier': 5
 }
 
+# Prior bounds (added log_sigma bound)
 PRIOR_BOUNDS = {
-    't0': (58000.0, 59000.0),
-    'x0': (-5.0, -2.6),
-    'x1': (-4.0, 4.0),
-    'c': (-0.3, 0.3),
-    'log_sigma': (-3.0, 1.0)
+    'z': {'min': 0.001, 'max': 0.2},
+    't0': {'min': 58000.0, 'max': 59000.0},
+    'x0': {'min': -5.0, 'max': -2.6},
+    'x1': {'min': -4.0, 'max': 4.0},
+    'c': {'min': -0.3, 'max': 0.3},
+    'log_sigma': {'min': -3.0, 'max': 1.0}
 }
 
 # Enable float64 precision
 jax.config.update("jax_enable_x64", True)
 
-# Load data with fixed redshift
-times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = \
-    load_and_process_data('19dwz', data_dir='data', fix_z=True)
+# Load and process data
+times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = load_and_process_data('19dwz', data_dir='data', fix_z=fix_z)
 
-z = fixed_z[0]
+# Create SALT3 source for bandflux calculations
 source = SALT3Source()
 
-# Setup priors
-base_params = ['t0', 'x0', 'x1', 'c']
-prior_params = base_params + (['log_sigma'] if fit_sigma else [])
-prior_dists = {p: distrax.Uniform(low=PRIOR_BOUNDS[p][0], high=PRIOR_BOUNDS[p][1])
-               for p in prior_params}
+# Define parameter bounds and priors
+if fix_z:
+    param_bounds = {
[…]
