### 2025-10-10 7c9fa5f
Simplify ns_anomaly.py: reduce from 1720 to 458 lines
Major simplifications:
- Remove complex YAML configuration (use simple variables at top)
- Remove custom data loading (use standard load_and_process_data)
- Remove custom bandpass handling (see custom_bandpass_example.py)
- Remove true value comparisons and x0_mag plotting
- Streamline to 3 essential plots: corner comparison, anomaly corner, light curve
- Keep all core anomaly detection functionality intact
- Use save_chains_dead_birth utility for consistent chain saving
- Fix shape handling in logprior/loglikelihood to match ns.py pattern

[…]
