+++ examples/ns_anomaly.py
@@ -1,14 +1,14 @@
 """
 Anomaly detection for supernova light curves using nested sampling.
 
-This script implements a Bayesian anomaly detection framework for supernova
-light curve analysis. It runs two nested sampling procedures:
-1. A standard version that fits SALT3 model parameters to the data
-2. An anomaly detection version that includes an additional parameter (log_p)
-   to identify potential outliers in the data
+This script implements a Bayesian anomaly detection framework that:
+1. Runs standard SALT3 nested sampling
+2. Runs anomaly detection nested sampling with an additional log_p parameter
+3. Identifies potential outlier data points using weighted emax values
+4. Compares the two approaches via corner plots and light curve fits
 
-The script generates comparison plots, parameter statistics, and weighted
-anomaly indicators (emax values) that highlight potential problematic data points.
+The anomaly detection adds a parameter log_p that allows individual data points
+to be down-weighted if they don't fit the model well.
 """
 import distrax
 import os
@@ -18,1703 +18,477 @@ import jax.numpy as jnp
 import numpy as np
 import tqdm
 import blackjax
-import yaml
-import pandas as pd
 from blackjax.ns.utils import log_weights
-from jax_supernovae.salt3 import optimized_salt3_multiband_flux, precompute_bandflux_bridge
-from jax_supernovae.bandpasses import register_bandpass, get_bandpass, register_all_bandpasses, Bandpass
+from jax_supernovae import SALT3Source
+from jax_supernovae.data import load_and_process_data
 from jax_supernovae.utils import save_chains_dead_birth
-from jax_supernovae.data import load_redshift, load_hsf_data
 import matplotlib.pyplot as plt
 from anesthetic import read_chains, make_2d_axes
-import requests
 
+# ============================================================================
+# Configuration
+# ============================================================================
 
-# Define default settings for nested sampling and prior bounds
-DEFAULT_NS_SETTINGS = {
-    'max_iterations': int(os.environ.get('NS_MAX_ITERATIONS', '500')),
-    'n_delete': 75,
-    'n_live': 150,
+# Supernova to analyze
+SN_NAME = '19dwz'
+
+# Nested sampling settings
+NS_SETTINGS = {
+    'n_delete': 60,
+    'n_live': 125,
     'num_mcmc_steps_multiplier': 5,
-    'fit_sigma': False,
-    'fit_log_p': True,
-    'fit_z': True
+    'max_iterations': 500
 }
 
-DEFAULT_PRIOR_BOUNDS = {
-    'z': {'min': 0.001, 'max': 0.2},
-    't0': {'min': 58000.0, 'max': 60000.0},
-    'x0': {'min': -5.0, 'max': -1},
-    'x1': {'min': -10, 'max': 10},
-    'c': {'min': -0.6, 'max': 0.6},
-    'sigma': {'min': 0.001, 'max': 5},
+# Prior bounds
+PRIOR_BOUNDS = {
+    't0': {'min': 58000.0, 'max': 59000.0},
+    'x0': {'min': -5.0, 'max': -2.6},
+    'x1': {'min': -4.0, 'max': 4.0},
+    'c': {'min': -0.3, 'max': 0.3},
     'log_p': {'min': -20, 'max': -1}
 }
 
-# Default settings
-DEFAULT_SETTINGS = {
-    'fix_z': True,
-    'sn_name': '19vnk',  # Default supernova to analyze
-    'selected_bandpasses': None,  # Default: use all available bandpasses
-    'custom_bandpass_files': None  # Default: no custom bandpass files
-}
-
-# Try to load settings.yaml; if not found, use an empty dictionary
-try:
[…]
