### 2025-10-10 03b14b3
Update v3.0 API: fix param_names, add method conversions, register Bessell filters, fix test_documentation.py
Core SALT3Source changes:
- param_names now returns ['x0', 'x1', 'c'] to match sncosmo
- Convert minphase/maxphase/minwave/maxwave from properties to methods
- Update __str__ to mention 'v3.0 functional API'
- Add sncosmo fallback for Bessell filters in get_bandpass()

test_documentation.py fixes:
- Update all load_and_process_data calls to use optimized path with bridges
- Use band_indices, bridges, unique_bands for efficient multi-band calculations
- Remove bands=unique_bands pattern
[…]
