+++ jax_supernovae/__init__.py
@@ -1,3 +1,4 @@
 from jax_supernovae.source import SALT3Source
+from jax_supernovae.data import load_and_process_data
 
-__all__ = ['SALT3Source']
+__all__ = ['SALT3Source', 'load_and_process_data']
