+++ parameter_statistics.txt
@@ -2,8 +2,8 @@ Parameter Statistics:
 --------------------------------------------------
 Parameter               Mean         Std Dev
 --------------------------------------------------
-t0              58606.512942        0.654155
-log_x0             -2.966406        0.025186
-x1                  1.433097        0.711414
-c                   0.048316        0.070168
+t0              58606.433556        0.574752
+log_x0             -2.976827        0.027840
+x1                  0.577633        0.570225
+c                   0.166789        0.065713
 --------------------------------------------------
