### 2025-10-11 721bd1b
Fix v3.0 functional API: scalar returns and parameter handling
- Fix scalar return type detection in salt3.py and source.py
  * Check if input is scalar BEFORE converting to arrays
  * Fixes test_api_compatibility.py scalar return tests

- Fix bandmag method signature to match sncosmo
  * Change parameter order to (params, bands, magsys, phases)
  * Rename zpsys to magsys for consistency

- Add support for array zpsys/magsys
  * Handle lists/arrays of magnitude systems
  * Fixes test_broadcast_with_array_zpsys

- Fix zeropoint handling when zp=None
  * Don't pass zeros array when no zeropoint 
[…]
