+++ jax_supernovae/source.py
@@ -145,6 +145,11 @@ class SALT3Source:
         if 'x0' not in params or 'x1' not in params or 'c' not in params:
             raise ValueError("params must contain 'x0', 'x1', and 'c'")
 
+        # Check if input is scalar BEFORE any conversions
+        scalar_phase_input = np.isscalar(phases)
+        scalar_band_input = isinstance(bands, str)
+        scalar_input = scalar_phase_input and scalar_band_input
+
         # Create full parameter dict with defaults
         full_params = {
             'z': params.get('z', 0.0),
@@ -198,9 +203,6 @@ class SALT3Source:
         # This is simpler but slower - fine for one-off calculations
         # but inefficient for nested sampling (use bridges parameter instead)
 
-        # Determine if input is scalar
-        scalar_input = isinstance(bands, str) and np.isscalar(phases)
-
         # Convert bands and phases to arrays
         if isinstance(bands, str):
             bands_arr = [bands]
@@ -218,7 +220,13 @@ class SALT3Source:
             else:
                 zps_arr = zps
         else:
-            zps_arr = jnp.zeros(len(phases_arr))
+            zps_arr = None
+
+        # Handle zpsys (can be scalar string or array of strings)
+        if isinstance(zpsys, (list, tuple, np.ndarray)):
+            zpsys_arr = list(zpsys)
+        else:
+            zpsys_arr = None  # Will use zpsys directly as scalar
 
         # If phases and bands have same length, calculate one flux per (phase, band) pair
         if len(bands_arr) == len(phases_arr):
@@ -226,9 +234,11 @@ class SALT3Source:
             for i, (band, phase) in enumerate(zip(bands_arr, phases_arr)):
                 bandpass = get_bandpass(band)
                 bridge = precompute_bandflux_bridge(bandpass)
+                curr_zp = zps_arr[i] if zps_arr is not None else None
+                curr_zpsys = zpsys_arr[i] if zpsys_arr is not None else zpsys
                 flux = optimized_salt3_bandflux(
                     phase, bridge['wave'], bridge['dwave'], bridge['trans'],
-                    full_params, zp=zps_arr[i], zpsys=zpsys
+                    full_params, zp=curr_zp, zpsys=curr_zpsys
                 )
                 fluxes.append(flux)
             result = np.array(fluxes)
@@ -240,9 +250,11 @@ class SALT3Source:
             bridge = precompute_bandflux_bridge(bandpass)
             fluxes = []
             for i, phase in enumerate(phases_arr):
+                curr_zp = zps_arr[i] if zps_arr is not None else None
+                curr_zpsys = zpsys_arr[i] if zpsys_arr is not None else zpsys
                 flux = optimized_salt3_bandflux(
                     phase, bridge['wave'], bridge['dwave'], bridge['trans'],
-                    full_params, zp=zps_arr[i], zpsys=zpsys
+                    full_params, zp=curr_zp, zpsys=curr_zpsys
                 )
                 fluxes.append(flux)
             result = np.array(fluxes)
@@ -255,9 +267,11 @@ class SALT3Source:
             for i, band in enumerate(bands_arr):
                 bandpass = get_bandpass(band)
                 bridge = precompute_bandflux_bridge(bandpass)
+                curr_zp = zps_arr[i] if zps_arr is not None else None
+                curr_zpsys = zpsys_arr[i] if zpsys_arr is not None else zpsys
                 flux = optimized_salt3_bandflux(
                     phase, bridge['wave'], bridge['dwave'], bridge['trans'],
-                    full_params, zp=zps_arr[i], zpsys=zpsys
+                    full_params, zp=curr_zp, zpsys=curr_zpsys
                 )
                 fluxes.append(flux)
             return np.array(fluxes)
[…]
