### 2025-10-11 19ed83a
Make SALT3Source fully JIT-compatible for nested sampling
The v3.0 functional API refactor's primary goal was JIT compatibility.
This commit completes that goal by removing all JIT-incompatible code.

## Key Changes

### 1. Return JAX Arrays Instead of NumPy (source.py)
- Changed `np.array()` → `jnp.stack()` for array results
- Changed `float(result[0])` → `result[0]` for scalar results
- Keeps return values as JAX arrays for JIT traceability
- Both optimized and standard paths now JIT-compatible

### 2. Pre-Register Bessell Bands (bandpasses.py)
- Added Bessell filter loading from sncosmo at 
[…]
