### 2025-10-11 dae10e0
Update all examples to use SALT3Source v3.0 functional API
- Migrate all 6 examples from low-level API to SALT3Source class
- Add @jax.jit decorators to nested sampling examples (ns_with_shifts, ns_multi_sn_shifts, fmin_bfgs)
- Add shifts parameter to source.py bandflux() for transmission shift support
- Update plotting examples (plot_light_curve, plot_comparison) to use source.bandflux()
- Simplify dust_extinction_example.py to use band names directly
- Remove obsolete ns_gpu.py example (functionality replaced by standard ns.py)

All examples now use v3.0 functional API pattern:
- Parameters 
[…]
