### 2025-10-13 eb428eb
Clarify BlackJAX nested sampling dependency throughout documentation
- Update CLAUDE.md, README.rst, setup.py, requirements.txt to specify Handley Lab fork
- Note that this fork is not yet merged with main BlackJAX branch
- Add IMPORTANT notes to all nested sampling example files (ns.py, ns_anomaly.py, ns_with_shifts.py, ns_multi_sn_shifts.py)
- Include installation instructions and link to nested sampling book
