### 2025-10-13 5d5038b
Fix Python version requirement to >=3.10 instead of >=3.11
