### 2025-10-13 4497e98
Fix band_indices comparison bug in plot_light_curve.py
- band_indices contains integers (0, 1, 2...), not band names
- Fixed loop to compare integers properly
- Updated labels to use actual band names from unique_bands
