### 2025-10-29 fb8b60b
Add Academic Use section to README with citation information
Added a new section instructing users to cite the JAX-bandflux paper (arXiv:2504.08081) when using the package in academic research.
