### 2025-11-03 ee0b669
Add TimeSeriesSource implementation for custom SED models
Implements TimeSeriesSource class matching sncosmo API while maintaining JAX/GPU compatibility. Enables fitting arbitrary spectral time series models with full numerical consistency (<0.01% difference with sncosmo).

Core Implementation:
- jax_supernovae/timeseries.py: JAX-compatible interpolation and bandflux calculation
  - interpolate_timeseries_2d: Bicubic interpolation using SALT3's kernval
  - timeseries_flux: Amplitude scaling and zero_before handling
  - timeseries_bandflux: Single-band integration
  - timeseries_multiband_flux:
[…]
