+++ tests/test_timeseries_source.py
@@ -1,0 +1,659 @@
+"""Test TimeSeriesSource API consistency with sncosmo.
+
+This module tests that the JAX-bandflux TimeSeriesSource implementation produces
+results exactly consistent with sncosmo's TimeSeriesSource. This is the CRITICAL
+test suite that verifies the functional API maintains numerical accuracy.
+
+Tolerance: rtol=1e-4 (0.01% difference) matching SALT3 tests
+"""
+
+import os
+import sys
+
+# Add project root to Python path if running the file directly
+if __name__ == "__main__":
+    project_root = os.path.abspath(os.path.join(os.path.dirname(__file__), '..'))
+    sys.path.insert(0, project_root)
+
+import numpy as np
+import jax
+import jax.numpy as jnp
+import sncosmo
+from jax_supernovae import TimeSeriesSource
+from jax_supernovae.bandpasses import get_bandpass
+from jax_supernovae.salt3 import precompute_bandflux_bridge
+
+
+# ============================================================================
+# Test Fixtures - Create Simple Test Models
+# ============================================================================
+
+def create_simple_gaussian_sed():
+    """Create simple Gaussian SED for testing.
+
+    Returns
+    -------
+    phase : np.ndarray
+        Phase grid (-20 to 50 days)
+    wave : np.ndarray
+        Wavelength grid (3000 to 9000 Angstroms)
+    flux : np.ndarray
+        2D flux array (Gaussian in time and wavelength)
+    """
+    phase = np.linspace(-20, 50, 100)
+    wave = np.linspace(3000, 9000, 200)
+
+    # Create 2D grids
+    p_grid, w_grid = np.meshgrid(phase, wave, indexing='ij')
+
+    # Gaussian in time (peak at phase=0, width=10 days)
+    time_profile = np.exp(-0.5 * (p_grid / 10.0)**2)
+
+    # Gaussian in wavelength (peak at 5000 Å, width=1000 Å)
+    wave_profile = np.exp(-0.5 * ((w_grid - 5000.0) / 1000.0)**2)
+
+    # Combined flux (scaled to realistic levels)
+    flux = time_profile * wave_profile * 1e-15
+
+    return phase, wave, flux
+
+
+def create_double_peaked_sed():
+    """Create SED with two peaks in time for testing.
+
+    Returns phase, wave, flux arrays.
+    """
+    phase = np.linspace(-30, 60, 120)
+    wave = np.linspace(2500, 10000, 250)
+
+    p_grid, w_grid = np.meshgrid(phase, wave, indexing='ij')
+
+    # Two Gaussians in time (blue and red peaks)
+    peak1 = np.exp(-0.5 * ((p_grid - 5.0) / 8.0)**2)  # Early blue peak
+    peak2 = np.exp(-0.5 * ((p_grid - 20.0) / 12.0)**2)  # Late red peak
+
+    # Wavelength dependence (blue vs red)
+    blue_profile = np.exp(-0.5 * ((w_grid - 4000.0) / 800.0)**2)
+    red_profile = np.exp(-0.5 * ((w_grid - 6500.0) / 1200.0)**2)
+
+    flux = (peak1 * blue_profile + peak2 * red_profile) * 5e-16
+
+    return phase, wave, flux
+
+
+# ============================================================================
+# Interpolation Tests
+# ============================================================================
+
+def test_interpolation_cubic():
+    """Test cubic interpolation matches sncosmo exactly.
+
+    Creates TimeSeriesSource with cubic interpolation (time_spline_degree=3)
[…]
