+++ examples/custom_sed_example.py
@@ -1,0 +1,464 @@
+"""Custom SED Example: Loading and Fitting Real Spectral Energy Distributions
+
+This script demonstrates how to create TimeSeriesSource models from real SED data:
+
+1. Using sncosmo's built-in spectral templates (Hsiao Type Ia template)
+2. Loading SED data from text files
+3. Creating custom SEDs from analytical models
+4. Fitting custom SEDs to photometric observations
+5. Comparing custom SEDs with SALT3
+
+Author: JAX-bandflux team
+Date: 2025
+"""
+
+import numpy as np
+import matplotlib.pyplot as plt
+from jax_supernovae import TimeSeriesSource, SALT3Source
+from jax_supernovae.bandpasses import get_bandpass
+from jax_supernovae.salt3 import precompute_bandflux_bridge
+import jax
+import jax.numpy as jnp
+
+# Try to import sncosmo for spectral templates
+try:
+    import sncosmo
+    HAS_SNCOSMO = True
+except ImportError:
+    HAS_SNCOSMO = False
+    print("Warning: sncosmo not available. Some examples will be skipped.")
+
+
+# ============================================================================
+# Example 1: Load Hsiao Type Ia Template from sncosmo
+# ============================================================================
+
+def example_1_hsiao_template():
+    """Create TimeSeriesSource from sncosmo's Hsiao Type Ia template."""
+    print("=" * 70)
+    print("Example 1: Loading Hsiao Type Ia Template")
+    print("=" * 70)
+
+    if not HAS_SNCOSMO:
+        print("Skipping - sncosmo not available\n")
+        return None
+
+    # Load Hsiao template from sncosmo
+    print("\nLoading Hsiao template from sncosmo...")
+    hsiao = sncosmo.get_source('hsiao')
+
+    # Define phase and wavelength grids
+    # Hsiao template is defined from -20 to +85 days
+    phases = np.arange(-20.0, 86.0, 1.0)  # 1-day spacing
+    wavelengths = np.arange(2000.0, 10000.0, 10.0)  # 10 Angstrom spacing
+
+    print(f"Phase range: [{phases[0]:.1f}, {phases[-1]:.1f}] days ({len(phases)} points)")
+    print(f"Wavelength range: [{wavelengths[0]:.0f}, {wavelengths[-1]:.0f} Å ({len(wavelengths)} points)")
+
+    # Create 2D flux grid by evaluating Hsiao at all phase/wavelength points
+    print("\nGenerating 2D flux grid...")
+    flux_grid = np.zeros((len(phases), len(wavelengths)))
+
+    for i, phase in enumerate(phases):
+        # Get flux spectrum at this phase
+        flux_grid[i, :] = hsiao._flux(phase, wavelengths)
+
+    print(f"Flux grid shape: {flux_grid.shape}")
+    print(f"Flux range: [{np.min(flux_grid):.3e}, {np.max(flux_grid):.3e}] erg/s/cm²/Å")
+
+    # Create TimeSeriesSource
+    print("\nCreating TimeSeriesSource from Hsiao template...")
+    source_jax = TimeSeriesSource(
+        phases, wavelengths, flux_grid,
+        zero_before=True,
+        time_spline_degree=3,
+        name='hsiao',
+        version='sncosmo'
+    )
+
+    print(f"✓ Created JAX TimeSeriesSource")
+    print(f"  Parameters: {source_jax.param_names}")
+    print(f"  Phase range: [{source_jax.minphase():.1f}, {source_jax.maxphase():.1f}] days")
+    print(f"  Wavelength range: [{source_jax.minwave():.0f}, {source_jax.maxwave():.0f}] Å")
+
[…]
