### 2025-11-04 f47df18
simplify examples
