+++ examples/ns.py
@@ -1,389 +1,159 @@
 """
-# Nested Sampling with JAX-bandflux
+Nested sampling example for SALT3 parameter inference using JAX-bandflux.
 
-This script demonstrates how to run the nested sampling procedure for supernovae SALT model fitting using the JAX-bandflux package. We will install the package, load the data, set up and run the nested sampling algorithm, and finally produce a corner plot of the posterior samples.
-
-IMPORTANT: This example requires the Handley Lab fork of BlackJAX (not yet merged with main branch).
-Install with: pip install git+https://github.com/handley-lab/blackjax@proposal
-See: https://handley-lab.co.uk/nested-sampling-book/intro.html
-
-For more examples and the complete codebase, visit the [JAX-bandflux GitHub repository](https://github.com/samleeney/JAX-bandflux). The academic paper associated with this work can be found [here](https://github.com/samleeney/JAX-bandflux/blob/71ca8d1b3b273147e1e9bf60a9ef11a806363b80/paper.bib).
+This variant only covers the configuration used in the accompanying paper
+and slides: the redshift is fixed to its measured heliocentric value and we
+introduce a hyper-parameter `log_sigma` that rescales the reported flux errors
+(`fit_sigma = True`). Removing the other branches keeps the example short
+and easier to follow.
 """
 
-
-import distrax
+import anesthetic
+import blackjax
 import jax
 import jax.numpy as jnp
-import numpy as np
-import tqdm
-import blackjax
-import os
-from blackjax.ns.utils import log_weights
-from jax_supernovae import SALT3Source
-from jax_supernovae.utils import save_chains_dead_birth
-from jax_supernovae.data import load_and_process_data
 import matplotlib.pyplot as plt
-from anesthetic import read_chains, make_2d_axes
+import numpy as np
+import os
+import tqdm
+from blackjax.ns.utils import finalise, uniform_prior
+from jax_supernovae import SALT3Source
+from jax_supernovae.data import load_and_process_data
 
-# Option to fit sigma or not.
-# When fit_sigma is True, the code will include an extra free parameter (log_sigma)
-# so that the effective flux error becomes sigma * flux_err with sigma = 10**(log_sigma).
-# When fit_sigma is False, the code will use the measured flux_err from the file.
-fit_sigma = False
-
-# Settings that were previously in YAML
-fix_z = True
-
-NS_SETTINGS = {
-    'n_delete': 60,
-    'n_live': 125,
-    'num_mcmc_steps_multiplier': 5
-}
-
-# Prior bounds (added log_sigma bound)
-PRIOR_BOUNDS = {
-    'z': {'min': 0.001, 'max': 0.2},
-    't0': {'min': 58000.0, 'max': 59000.0},
-    'x0': {'min': -5.0, 'max': -2.6},
-    'x1': {'min': -4.0, 'max': 4.0},
-    'c': {'min': -0.3, 'max': 0.3},
-    'log_sigma': {'min': -3.0, 'max': 1.0}
-}
-
-# Enable float64 precision
+# Enable float64 precision for the model and nested sampling machinery.
 jax.config.update("jax_enable_x64", True)
 
-# Load and process data
-times, fluxes, fluxerrs, zps, band_indices, unique_bands, bridges, fixed_z = load_and_process_data('19dwz', data_dir='data', fix_z=fix_z)
+# Configuration constants
+SUPERNOVA_ID = "19dwz"
+DATA_DIR = "data"
+FIXED_Z = 0.04607963148708845
[…]
