+++ examples/ns_anomaly.py
@@ -1,498 +1,457 @@
 """
-Anomaly detection for supernova light curves using nested sampling.
+Nested sampling anomaly detection example using SALT3 with structured priors.
 
-This script implements a Bayesian anomaly detection framework that:
-1. Runs standard SALT3 nested sampling
-2. Runs anomaly detection nested sampling with an additional log_p parameter
-3. Identifies potential outlier data points using weighted emax values
-4. Compares the two approaches via corner plots and light curve fits
+This script mirrors the ``ns.py`` configuration style:
+    - Parameters live in dictionaries compatible with JAX PyTrees
+    - Uniform priors are generated via ``blackjax.ns.utils.uniform_prior``
+    - Log-likelihoods are JIT-compiled and operate on structured inputs
 
-The anomaly detection adds a parameter log_p that allows individual data points
-to be down-weighted if they don't fit the model well.
-
-IMPORTANT: This example requires the Handley Lab fork of BlackJAX (not yet merged with main branch).
-Install with: pip install git+https://github.com/handley-lab/blackjax@proposal
-See: https://handley-lab.co.uk/nested-sampling-book/intro.html
+On top of the standard SALT3 fit it includes an anomaly model with a ``log_p``
+parameter that can down-weight outlying photometric points. Both runs are
+compared via corner plots and a light-curve diagnostic that highlights likely
+anomalies.
 """
-import distrax
+
 import os
-os.environ['XLA_PYTHON_CLIENT_PREALLOCATE'] = 'false'
+import anesthetic
+import blackjax
 import jax
 import jax.numpy as jnp
+import matplotlib.pyplot as plt
 import numpy as np
 import tqdm
-import blackjax
-from blackjax.ns.utils import log_weights
+from anesthetic import make_2d_axes
+from blackjax.ns.utils import finalise, uniform_prior, log_weights
+from jax.scipy.special import logsumexp
 from jax_supernovae import SALT3Source
 from jax_supernovae.data import load_and_process_data
-from jax_supernovae.utils import save_chains_dead_birth
-import matplotlib.pyplot as plt
-from anesthetic import read_chains, make_2d_axes
 
-# ============================================================================
-# Configuration
-# ============================================================================
-
-# Supernova to analyze
-SN_NAME = '19dwz'
-
-# Nested sampling settings
-NS_SETTINGS = {
-    'n_delete': 60,
-    'n_live': 125,
-    'num_mcmc_steps_multiplier': 5,
-    'max_iterations': 500
-}
-
-# Prior bounds
-PRIOR_BOUNDS = {
-    't0': {'min': 58000.0, 'max': 59000.0},
-    'x0': {'min': -5.0, 'max': -2.6},
-    'x1': {'min': -4.0, 'max': 4.0},
-    'c': {'min': -0.3, 'max': 0.3},
-    'log_p': {'min': -20, 'max': -1}
-}
-
-# Whether to fix redshift (True = use spectroscopic z)
-FIX_Z = True
-
-# Enable float64 precision
+# Enable float64 precision for the model and nested sampling machinery.
 jax.config.update("jax_enable_x64", True)
 
-# ============================================================================
-# Load data
-# ============================================================================
+# Configuration constants
+SUPERNOVA_ID = "19dwz"
[…]
