### 2025-11-14 #8150 @JBorrow
Finally (apologies for the spam, all), @samleeney, please note that we do have a new interim policy on AI-assisted submissions: https://joss.readthedocs.io/en/latest/submitting.html#interim-position-on-ai-enabled-software. The EiC of the track and I have reviewed your repository, and are happy that it is not in violation with this policy and that the review can go ahead. Thank you also for ensuring that AI-generated commits are tagged as being authored by Claude.
