### 2025-11-14 #8150 @Samreay
Im happy to help with the review too (on Monday haha)! It'll actually be my first JOSS review to give back, so will read over the process [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) :)
