### 2025-11-17 #8150 @wbenoit26
Yes, I can review this paper
