### 2025-11-17 #9441 @editorialbot
[REVIEW]: JAX-bandflux: differentiable supernovae SALT modelling for cosmological analysis on GPUs
**Submitting author:** @samleeney (<a href="http://orcid.org/0000-0003-4366-1119">Samuel Alan Kossoff Leeney</a>)
**Repository:** https://github.com/samleeney/JAX-bandflux
**Branch with paper.md** (empty if default branch):
**Version:** 1.0.0
