### 2025-11-18 #9441 @JBorrow
Thanks @Samreay; it is not out of scope to request changes to improve packaging, especially in situations where you get different results depending on how you install a package. It is reasonable to ask for a change to `pyproject.toml` to enhance compatibility with modern tooling like `uv` and to reduce duplication, though if the author can use a different technique to accommodate that then we should allow it.
