### 2025-11-19 1656430
Fix quickstart installation and add CI test
Addresses JOSS reviewer feedback on installation issues:

- Fix JAX/JAXlib version compatibility in setup.py by removing explicit
  jaxlib dependency (JAX installs matching version automatically)
- Update quickstart instructions to install from GitHub instead of PyPI
  to ensure latest features (SALT3Source, TimeSeriesSource) are available
- Add CI job to test quickstart example runs successfully
- Add guidance for CUDA/GPU installation

Fixes #4

🤖 Generated with [Claude Code](https://claude.com/claude-code)

Co-Authored-By: Claude <noreply@anthropi
[…]
