### 2025-12-12 0c1b954
Add data format guidance and example template
