### 2025-12-12 6003e9f
Expose data namespace and align load_and_process_data examples
