+++ jax_supernovae/__init__.py
@@ -1,4 +1,5 @@
 from jax_supernovae.source import SALT3Source, TimeSeriesSource
 from jax_supernovae.data import load_and_process_data
+from . import data
 
-__all__ = ['SALT3Source', 'TimeSeriesSource', 'load_and_process_data']
+__all__ = ['SALT3Source', 'TimeSeriesSource', 'load_and_process_data', 'data']
