### 2025-12-12 cf89f68
Allow Bandpass name parameter and document it
