### 2025-12-12 1a4a320
Consolidate packaging and clarify CPU/GPU install paths
