+++ docs/requirements.txt
@@ -1,4 +1,1 @@
-sphinx>=4.0.0
-sphinx-rtd-theme>=1.0.0
-sphinxcontrib-mermaid>=0.7.0
-sphinx-autodoc-typehints>=1.12.0
+.[docs]
