+++ .github/workflows/test.yml
@@ -21,14 +21,7 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install -e .
-        pip install pytest numpy jax sncosmo pyyaml matplotlib
-        pip install git+https://github.com/handley-lab/blackjax@proposal
-        pip install git+https://github.com/google-deepmind/distrax
-        pip install tqdm
-        pip install -r requirements.txt
-        # Install documentation-specific dependencies
-        pip install -r docs/requirements.txt
+        pip install -e .[dev,nested,docs]
 
     - name: Run tests
       run: |
