+++ pyproject.toml
@@ -1,3 +1,68 @@
 [build-system]
 requires = ["setuptools>=61.0", "wheel"]
-build-backend = "setuptools.build_meta" 
+build-backend = "setuptools.build_meta"
+
+[project]
+name = "jax-bandflux"
+version = "0.3.3"
+description = "JAX-based supernova bandflux modelling with SALT3"
+readme = "README.md"
+requires-python = ">=3.10"
+license = { text = "MIT" }
+authors = [
+  { name = "Samuel Alan Kossoff Leeney", email = "sakl2@cam.ac.uk" }
+]
+keywords = ["supernova", "jax", "sncosmo", "astronomy", "salt3"]
+classifiers = [
+  "Development Status :: 3 - Alpha",
+  "Intended Audience :: Science/Research",
+  "License :: OSI Approved :: MIT License",
+  "Programming Language :: Python :: 3",
+  "Programming Language :: Python :: 3 :: Only",
+  "Programming Language :: Python :: 3.10",
+  "Topic :: Scientific/Engineering :: Astronomy",
+]
+dependencies = [
+  "numpy>=1.24.0",
+  "jax>=0.4.20",
+  "sncosmo>=2.9.0",
+  "astropy>=5.0.0",
+  "pyyaml>=6.0.0",
+  "matplotlib",
+  "tqdm",
+  "requests",
+]
+
+[project.optional-dependencies]
+docs = [
+  "sphinx>=5.0.0",
+  "sphinx-rtd-theme>=1.0.0",
+  "sphinxcontrib-mermaid>=0.7.0",
+  "sphinx-autodoc-typehints>=1.12.0",
+]
+dev = [
+  "pytest>=7.0",
+  "pytest-cov>=2.0",
+  "black>=22.0",
+  "isort>=5.0",
+]
+nested = [
+  "anesthetic",
+  "blackjax @ git+https://github.com/handley-lab/blackjax.git",
+  "distrax",
+]
+cuda12 = [
+  "jax[cuda12]>=0.4.20",
+]
+
+[tool.setuptools]
+packages = { find = { include = ["jax_supernovae"] } }
+
+[tool.setuptools.package-data]
+jax_supernovae = [
+  "data/**/*",
+  "sncosmo-modelfiles/**/*",
+]
+
+[tool.setuptools.exclude-package-data]
+"*" = ["__pycache__/*", "*.py[cod]"]
