### 2025-12-12 924ddb3
Add CONTRIBUTING and surface support info
