### 2025-12-12 e2d4ceb
Tune performance benchmarks for CPU and add JIT
