### 2025-12-19 c597c7c
Use optimized SALT3 bandflux path in performance test
