+++ jax_supernovae/source.py
@@ -11,7 +11,7 @@ rather than stored in the source object.
 
 import jax.numpy as jnp
 import numpy as np
-from jax_supernovae.salt3 import optimized_salt3_multiband_flux, optimized_salt3_bandflux, precompute_bandflux_bridge
+from jax_supernovae.salt3 import optimized_salt3_multiband_flux, precompute_bandflux_bridge
 from jax_supernovae.bandpasses import register_all_bandpasses, get_bandpass
 from jax_supernovae.timeseries import (
     timeseries_bandflux,
@@ -66,6 +66,8 @@ class SALT3Source:
 
         # Register all bandpasses to ensure they're available
         register_all_bandpasses()
+        # Cache of precomputed bridges keyed by band name to avoid slow rebuilding
+        self._bridge_cache = {}
 
     @property
     def param_names(self):
@@ -101,9 +103,18 @@ class SALT3Source:
         """Official string representation."""
         return f"SALT3Source(name='{self.name}')"
 
+    def _get_bridges(self, unique_bands):
+        """Return cached bridges for the given bands, computing missing ones."""
+        bridges = []
+        for band in unique_bands:
+            if band not in self._bridge_cache:
+                self._bridge_cache[band] = precompute_bandflux_bridge(get_bandpass(band))
+            bridges.append(self._bridge_cache[band])
+        return tuple(bridges)
+
     def bandflux(self, params, bands, phases, zp=None, zpsys=None,
                  band_indices=None, bridges=None, unique_bands=None, shifts=None):
-        """Calculate bandflux using v3.0 functional API.
+        """Calculate bandflux using the optimized multiband path only.
 
         Parameters
         ----------
@@ -117,50 +128,34 @@ class SALT3Source:
             - 't0': float - Time of peak brightness (default: 0.0)
         bands : str or array-like
             Bandpass name(s). Can be a string for a single band or array of band names.
+            If `band_indices`/`unique_bands` are provided, `bands` can be None.
         phases : float or array
             Rest-frame phase(s) relative to t0.
         zp : float or array, optional
-            Zero point(s) for flux scaling
+            Zero point(s) for flux scaling (per observation).
         zpsys : str, optional
-            Zero point system ('ab' or 'vega')
+            Zero point system ('ab' currently supported).
         band_indices : array, optional
-            For performance: indices into unique_bands/bridges arrays.
-            If provided, must also provide bridges and unique_bands.
+            Indices into unique_bands/bridges arrays (for high-performance path).
         bridges : tuple, optional
-            For performance: precomputed bridge data structures.
+            Precomputed bridge data structures keyed to unique_bands.
         unique_bands : list, optional
-            For performance: list of unique band names corresponding to bridges.
+            List of unique band names corresponding to bridges.
         shifts : array or list, optional
-            Wavelength shifts in Angstroms for each unique band. Used to model
-            transmission uncertainties or systematic wavelength calibration errors.
+            Wavelength shifts in Angstroms for each unique band.
 
         Returns
         -------
         flux : array
-            Bandflux value(s) with shape matching the input
-
-        Examples
-        --------
-        Single band and phase:
-        >>> source = SALT3Source()
-        >>> params = {'x0': 1e-5, 'x1': 0.0, 'c': 0.0}
-        >>> flux = source.bandflux(params, 'bessellb', 0.0, zp=27.5, zpsys='ab')
-
-        High-performance mode with precomputed bridges:
-        >>> flux = source.bandflux(params, bands, phases, zp=zps, zpsys='ab',
-        ...                        band_indices=band_indices, bridges=bridges,
[…]
