+++ jax_supernovae/salt3.py
@@ -577,18 +577,23 @@ def precompute_bandflux_bridge(bandpass):
         - 'trans': the transmission values computed on the grid
         - 'wave_original': original wavelength array for shift interpolation
         - 'trans_original': original transmission array
+        - 'zpbandflux_ab': AB zeropoint normalization for this band
     """
     wave = bandpass.integration_wave
     dwave = bandpass.integration_spacing
     trans = bandpass(wave)
     
+    # Zeropoint normalization is constant per bandpass (AB system)
+    zpbandflux_ab = 3631e-23 * dwave / H_ERG_S * jnp.sum(trans / wave)
+
     # Store original arrays for shift interpolation
     return {
         'wave': wave, 
         'dwave': dwave, 
         'trans': trans,
         'wave_original': bandpass.wave,
-        'trans_original': bandpass.trans
+        'trans_original': bandpass.trans,
+        'zpbandflux_ab': zpbandflux_ab,
     }
 
 @jax.jit
