### 2025-12-19 6b2019d
Cache jitted SALT3 bandflux core to cut per-call overhead
