### 2025-12-19 183b30c
Update perf tests and docs for batched GPU speedups
The main issue was the test pointing to an old legacy function which is no longer relevant. Once it points to the correct function when evaluating *single* sets of parameters and *single* bandpasses we are approx equal to sncosmo, depending on which gpu you use.

However, the above case is not really how jax bandflux shoudl be used. when integrated into gpu based samplers using jax, the entire process will be jitted and fused together allowing the sampler to evaluate many sets of parameters simultaneously and furthermore allow fast running of
[…]
