+++ paper.bib
@@ -35,3 +35,10 @@
     year = {2022},
     publisher = {IOP Publishing}
 }
+
+@article{leeney2025anomaly,
+    title = {Bayesian Anomaly Detection for Ia Cosmology: Automating SALT3 Data Curation},
+    author = {Leeney, Samuel Alan Kossoff and Handley, William James and Bevins, H. T. J. and Acedo, E.},
+    journal = {arXiv preprint arXiv:2509.13394},
+    year = {2025}
+}
