+++ tests/test_bandflux_performance.py
@@ -15,6 +15,84 @@ from jax_supernovae.bandpasses import get_bandpass
 from jax_supernovae.salt3 import precompute_bandflux_bridge
 
 
+def test_multiband_batched_performance():
+    """Compare batched multi-band performance between JAX and sncosmo."""
+    params = {'x0': 1e-5, 'x1': -0.5, 'c': 0.2}
+    batch_size = 1000
+
+    # Create JAX source (v3.0 - no parameter storage)
+    jax_source = SALT3Source()
+
+    # Create and configure sncosmo source (stateful API for comparison)
+    snc_source = sncosmo.get_source('salt3-nir')
+
+    # Simulate a realistic light curve: 100 observations across 5 bands
+    np.random.seed(123)
+    n_obs = 100
+    bands = np.random.choice(['bessellb', 'bessellv', 'bessellr', 'besselli', 'bessellux'], n_obs)
+    unique_bands = list(sorted(set(bands)))
+    band_to_idx = {b: i for i, b in enumerate(unique_bands)}
+    band_indices = np.array([band_to_idx[b] for b in bands])
+    bridges = tuple(precompute_bandflux_bridge(get_bandpass(b)) for b in unique_bands)
+    phases = np.random.uniform(-10, 40, n_obs)
+    zp = 27.5
+    zpsys = 'ab'
+
+    # Build batched params
+    params_batch = {
+        'x0': np.random.uniform(1e-6, 1e-4, batch_size),
+        'x1': np.random.uniform(-2, 2, batch_size),
+        'c': np.random.uniform(-0.3, 0.3, batch_size),
+        'z': np.zeros(batch_size),
+        't0': np.zeros(batch_size),
+    }
+
+    # Warmup JAX batched path
+    print("\nWarming up JAX for batched multi-band calculations...")
+    _ = jax_source.bandflux_batch(
+        params_batch, bands, phases, zp=zp, zpsys=zpsys,
+        band_indices=band_indices, bridges=bridges, unique_bands=unique_bands
+    )
+
+    # Benchmark sncosmo (loop over batch)
+    print("Benchmarking sncosmo (batched params, multi-band)...")
+    start_time = time.time()
+    snc_fluxes = []
+    for i in range(batch_size):
+        snc_source.set(x0=params_batch['x0'][i],
+                       x1=params_batch['x1'][i],
+                       c=params_batch['c'][i])
+        flux = snc_source.bandflux(bands, phases, zp=zp, zpsys=zpsys)
+        snc_fluxes.append(flux)
+    snc_time = time.time() - start_time
+
+    # Benchmark JAX batched
+    print("Benchmarking JAX (batched params, multi-band)...")
+    start_time = time.time()
+    jax_fluxes = jax_source.bandflux_batch(
+        params_batch, bands, phases, zp=zp, zpsys=zpsys,
+        band_indices=band_indices, bridges=bridges, unique_bands=unique_bands
+    )
+    jax_time = time.time() - start_time
+
+    # Print results
+    print("\n" + "=" * 70)
+    print("Batched Multi-Band Performance Comparison")
+    print("=" * 70)
+    print(f"Batch size: {batch_size}, Observations per param: {n_obs}")
+    print(f"SNCosmo time:  {snc_time:.4f} seconds ({snc_time/batch_size*1000:.2f} ms/param)")
+    print(f"JAX time:      {jax_time:.4f} seconds ({jax_time/batch_size*1000:.2f} ms/param)")
+    print(f"Speedup:       {snc_time/jax_time:.2f}x")
+    print("=" * 70)
+
+    # Verify results match
+    snc_fluxes = np.array(snc_fluxes)
+    jax_fluxes = np.array(jax_fluxes)
[…]
