### 2025-12-19 8a40269
Handle array zpsys in bandflux cache key
