### 2025-12-19 f7fa4a1
Clarify CPU vs CUDA JAX install guidance
