### 2025-12-19 ca97635
Add contributing/support guidance
