### 2025-12-19 b4d451e
Add contributing/support guidance
