### 2025-12-19 79d0cd2
Add Sphinx autodoc API reference
