### 2025-12-19 09216bc
Add docs badge to README
