### 2025-12-20 5673911
Document integration merge closure plan
