### 2026-01-07 cebfb25
Add auto-publish workflow on push to master/main
