+++ .github/workflows/publish.yml
@@ -1,6 +1,8 @@
 name: Publish to PyPI
 
 on:
+  push:
+    branches: [ master, main ]
   release:
     types: [created]
 
@@ -11,27 +13,79 @@ jobs:
       name: pypi
       url: https://pypi.org/p/jax-bandflux
     permissions:
+      contents: write
       id-token: write
     steps:
     - uses: actions/checkout@v3
-    
+      with:
+        fetch-depth: 0
+        token: ${{ secrets.GITHUB_TOKEN }}
+
     - name: Set up Python
       uses: actions/setup-python@v4
       with:
         python-version: '3.11'
-    
+
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        pip install build twine
-    
+        pip install build twine toml
+        pip install -e .[dev,nested]
+
+    - name: Run tests
+      run: |
+        pytest tests/ -v
+
+    - name: Auto-increment version
+      id: version
+      run: |
+        python << 'EOF'
+        import toml
+        import os
+
+        # Read current version
+        with open('pyproject.toml', 'r') as f:
+            data = toml.load(f)
+
+        current_version = data['project']['version']
+        major, minor, patch = map(int, current_version.split('.'))
+
+        # Increment patch version
+        new_version = f"{major}.{minor}.{patch + 1}"
+
+        # Update version in pyproject.toml
+        data['project']['version'] = new_version
+        with open('pyproject.toml', 'w') as f:
+            toml.dump(data, f)
+
+        print(f"Updated version from {current_version} to {new_version}")
+
+        # Set output for later steps
+        with open(os.environ['GITHUB_OUTPUT'], 'a') as f:
+            f.write(f"new_version={new_version}\n")
+            f.write(f"old_version={current_version}\n")
+        EOF
+
+    - name: Commit version bump
+      run: |
+        git config --local user.email "github-actions[bot]@users.noreply.github.com"
+        git config --local user.name "github-actions[bot]"
+        git add pyproject.toml
+        git commit -m "Bump version to ${{ steps.version.outputs.new_version }} [skip ci]"
+        git push
+
     - name: Build package
       run: python -m build
-    
+
     - name: Publish to PyPI
       env:
         TWINE_USERNAME: __token__
         TWINE_PASSWORD: ${{ secrets.pypi2 }}
       run: |
         twine check dist/*
-        twine upload --verbose dist/* 
+        twine upload --verbose dist/*
+
+    - name: Create git tag
+      run: |
+        git tag "v${{ steps.version.outputs.new_version }}"
+        git push origin "v${{ steps.version.outputs.new_version }}" 
