### 2026-01-07 0f2d23d
Bump version to 0.3.4 [skip ci]
