### 2026-01-07 8eb7cbf
Fix blackjax dependency for PyPI compatibility
