+++ pyproject.toml
@@ -21,8 +21,11 @@ text = "MIT"
 [project.optional-dependencies]
 docs = [ "sphinx>=5.0.0", "sphinx-rtd-theme>=1.0.0", "sphinxcontrib-mermaid>=0.7.0", "sphinx-autodoc-typehints>=1.12.0",]
 dev = [ "pytest>=7.0", "pytest-cov>=2.0", "black>=22.0", "isort>=5.0",]
-nested = [ "anesthetic", "blackjax @ git+https://github.com/handley-lab/blackjax.git", "distrax",]
+nested = [ "anesthetic", "blackjax>=1.0.0", "distrax",]
 cuda12 = [ "jax[cuda12]>=0.4.20",]
+
+[tool.setuptools]
+packages.find.include = [ "jax_supernovae",]
 
 [tool.setuptools.package-data]
 jax_supernovae = [ "data/**/*", "sncosmo-modelfiles/**/*",]
