### 2026-01-07 59f54e4
Fix pyproject.toml syntax error - remove duplicate packages.find
