### 2026-01-07 f44305a
Bump version to 0.3.5 [skip ci]
