### 2026-01-08 edeed00
Read docs version dynamically from pyproject.toml
Previously the version was hardcoded in docs/conf.py and would get
out of sync with the package version. Now it reads from pyproject.toml
so documentation always shows the correct version.
