### 2026-01-08 709a889
Bump version to 0.3.6 [skip ci]
