### 2026-01-08 ed5670f
Rewrite documentation with doctest-validated examples and plots
- Replace placeholder outputs with actual computed values verified by sphinx.ext.doctest
- Add matplotlib plot_directive for interactive light curve visualizations
- Convert timeseries_source from markdown to RST format
- Add GitHub Actions workflow for automated documentation testing
- Fix SALT3 parameter description to match literature (3 parameters: x0, x1, c)
- All 278 doctests pass automatically during Sphinx build

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
