+++ docs/timeseries_source.md
@@ -1,524 +1,0 @@
-# TimeSeriesSource Documentation
-
-## Overview
-
-`TimeSeriesSource` is a JAX-bandflux class for fitting custom supernova spectral energy distributions (SEDs). It provides a JAX/GPU-accelerated implementation matching sncosmo's `TimeSeriesSource` API whilst using a functional parameter-passing approach for optimal performance in MCMC and nested sampling applications.
-
-## Key Features
-
-- **Custom SED Models**: Fit any spectral time series defined on a 2D (phase × wavelength) grid
-- **Bicubic Interpolation**: Matches sncosmo exactly using JAX primitives
-- **Functional API**: Parameters passed as dictionaries for JAX compatibility
-- **Two-Tier Performance**: Simple mode for convenience, optimised mode for speed
-- **JIT Compatible**: Works seamlessly in JIT-compiled likelihood functions
-- **GPU Accelerated**: Runs efficiently on GPUs via JAX
-- **Numerical Accuracy**: Matches sncosmo to <0.01% (tested)
-
-## API Comparison: sncosmo vs JAX-bandflux
-
-### Constructor (Nearly Identical)
-
-**sncosmo:**
-```python
-source = sncosmo.TimeSeriesSource(phase, wave, flux,
-                                   zero_before=False,
-                                   time_spline_degree=3,
-                                   name=None, version=None)
-```
-
-**JAX-bandflux:**
-```python
-source = TimeSeriesSource(phase, wave, flux,  # Same signature!
-                          zero_before=False,
-                          time_spline_degree=3,
-                          name=None, version=None)
-```
-
-### Method Calls (Functional API)
-
-**sncosmo (stateful):**
-```python
-source.set(amplitude=1.0)
-flux = source.bandflux('bessellb', 0.0, zp=25.0, zpsys='ab')
-```
-
-**JAX-bandflux (functional):**
-```python
-params = {'amplitude': 1.0}
-flux = source.bandflux(params, 'bessellb', 0.0, zp=25.0, zpsys='ab')
-```
-
-The key difference: JAX-bandflux passes parameters as a dictionary to each method call, enabling JAX to trace parameter dependencies for autodiff and JIT compilation.
-
-## Basic Usage
-
-### Creating a TimeSeriesSource
-
-```python
-import numpy as np
-from jax_supernovae import TimeSeriesSource
-
-# Define your model grid
-phase = np.linspace(-20, 50, 100)  # Days
-wave = np.linspace(3000, 9000, 200)  # Angstroms
-
-# Create flux array (phase × wavelength)
-# This could come from:
-# - Theoretical models
-# - Spectroscopic observations
-# - PCA components
-# - Machine learning models
-p_grid, w_grid = np.meshgrid(phase, wave, indexing='ij')
-flux = ...  # Your 2D flux array (erg/s/cm²/Å)
-
-# Create source
-source = TimeSeriesSource(phase, wave, flux,
-                          zero_before=False,  # Extrapolate before minphase
-                          time_spline_degree=3,  # Cubic interpolation
-                          name='my_model')
-```
-
-### Simple Photometry
-
-```python
-# Define parameters
-params = {'amplitude': 1.0}
-
-# Single observation
-flux_b = source.bandflux(params, 'bessellb', 0.0, zp=25.0, zpsys='ab')
-
-# Light curve (multiple phases, same band)
-phases = np.linspace(-10, 30, 50)
-fluxes_b = source.bandflux(params, 'bessellb', phases, zp=25.0, zpsys='ab')
-
[…]
