+++ docs/bandpass_loading.rst
@@ -1,62 +1,288 @@
 Bandpass Loading
-==============
+================
 
-What is a Bandpass?
-----------------
-A bandpass represents the transmission function of an astronomical filter, defining how much light at each wavelength passes through the filter. In supernova light curve modeling, bandpasses are essential for calculating the observed flux in different filters.
+JAX-bandflux provides flexible bandpass management for working with
+astronomical filters from various sources.
 
-Creating a Synthetic Bandpass
+Built-in Bandpasses
+-------------------
+
+JAX-bandflux includes common astronomical filters:
+
+.. doctest::
+
+   >>> # Access a built-in bandpass
+   >>> bp_b = get_bandpass('bessellb')
+   >>> print(f"Bessell B: {bp_b.minwave():.0f} - {bp_b.maxwave():.0f} Angstroms")
+   Bessell B: 3600 - 5600 Angstroms
+   >>>
+   >>> # Compare wavelength coverage across filters
+   >>> for name in ['bessellb', 'bessellv', 'bessellr', 'besselli']:
+   ...     bp = get_bandpass(name)
+   ...     print(f"  {name:10s}: {bp.minwave():7.1f} - {bp.maxwave():7.1f} A")
+     bessellb  :  3600.0 -  5600.0 A
+     bessellv  :  4700.0 -  7000.0 A
+     bessellr  :  5500.0 -  9000.0 A
+     besselli  :  7000.0 -  9200.0 A
+
+Available built-in bandpasses:
+
+- **Bessell**: ``bessellb``, ``bessellv``, ``bessellr``, ``besselli``, ``bessellux``
+- **SDSS**: ``g``, ``r``, ``i``, ``z``
+- **ZTF**: ``ztfg``, ``ztfr``
+- **ATLAS**: ``c``, ``o``
+- **2MASS**: ``H``
+
+Creating Custom Bandpasses
 --------------------------
+
+Create a bandpass from wavelength and transmission arrays:
+
+.. doctest::
+
+   >>> # Define a Gaussian bandpass
+   >>> wavelengths = np.linspace(4000, 5000, 100)
+   >>> transmission = np.exp(-((wavelengths - 4500) / 200)**2)
+   >>>
+   >>> # Create Bandpass object
+   >>> bandpass = Bandpass(wavelengths, transmission, name='custom_g')
+   >>> print(f"Custom bandpass '{bandpass.name}':")
+   Custom bandpass 'custom_g':
+   >>> print(f"  Wavelength range: {bandpass.minwave():.1f} - {bandpass.maxwave():.1f} A")
+     Wavelength range: 4000.0 - 5000.0 A
+   >>> print(f"  Peak wavelength: ~4500 A (Gaussian center)")
+   Peak wavelength: ~4500 A (Gaussian center)
+   >>> print(f"  Number of points: {len(wavelengths)}")
+     Number of points: 100
+
+The ``Bandpass`` class automatically:
+
+- Normalizes transmission values
+- Creates integration grids for flux calculations
+- Provides interpolation for arbitrary wavelengths
+
+Registering Bandpasses
+~~~~~~~~~~~~~~~~~~~~~~
+
+Register custom bandpasses for use by name:
+
+.. doctest::
+
+   >>> # Register the bandpass
+   >>> register_bandpass('my_filter', bandpass, force=True)
+   >>>
+   >>> # Now accessible by name
+   >>> retrieved = get_bandpass('my_filter')
+   >>> print(f"Retrieved '{retrieved.name}' as 'my_filter'")
+   Retrieved 'custom_g' as 'my_filter'
+   >>> print(f"Range: {retrieved.minwave():.1f} - {retrieved.maxwave():.1f} A")
+   Range: 4000.0 - 5000.0 A
+
+Loading from Files
[…]
