### 2026-01-08 39f8764
Bump version to 0.3.7 [skip ci]
