### 2026-01-08 44789f6
Bump version to 0.3.8 [skip ci]
