### 2026-01-08 91d9158
Fix docstring formatting for proper Sphinx rendering
- Convert doctest blocks (>>>) to testcode/testoutput format in RST files
- Convert Python docstring examples from >>> format to :: code blocks
- All 47 documentation tests still pass
- Code examples now render with proper syntax highlighting without >>> prompts

Fixes documentation rendering issues reported in JOSS review.

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
