### 2026-01-08 ae9efb1
Bump version to 0.3.9 [skip ci]
