+++ REVIEW_PR_MAPPING.md
@@ -1,27 +1,0 @@
-# Reviewer Issue-to-PR Mapping
-
-All PRs listed below are merged into `joss-review/integration` and should be closed once that branch is merged into `master`. Keep these links as the audit trail for reviewer re-checks.
-
-- **Issue 1: Data sourcing guidance** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/6
-
-- **Issue 2/3: Top-level data import & quickstart return shape** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/7
-
-- **Issue 4: Bandpass constructor name arg** — @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/8
-
-- **Issues 5–7: Install/docs dependency drift, SALT3 availability, CUDA backend/packaging alignment** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/9 (docs/installs), https://github.com/samleeney/JAX-bandflux/pull/14 (requirements aligned, no forced CUDA wheel)
-
-- **Issue 8: API reference/autodoc** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/13 (autodoc API page linked in docs)
-
-- **Issue 9: Community guidelines/support prominence** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/13 (CONTRIBUTING/support)
-
-- **Issue 10: Performance regression vs sncosmo** — @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/12 (perf tests on optimized path, batched multi-band benchmark showing ~100× GPU speedup per param; docs/paper note batched usage)
-
-- **Runtime import/pytest dep fix** — @Samreay, @wbenoit26  
-  PR: https://github.com/samleeney/JAX-bandflux/pull/10
