+++ requirements.txt
@@ -1,2 +1,0 @@
-# Development install uses the unified dependency definitions in pyproject.toml
-.[dev,nested]
