+++ parameter_statistics.txt
@@ -1,9 +1,0 @@
-Parameter Statistics:
---------------------------------------------------
-Parameter               Mean         Std Dev
---------------------------------------------------
-t0              58606.433556        0.574752
-log_x0             -2.976827        0.027840
-x1                  0.577633        0.570225
-c                   0.166789        0.065713
---------------------------------------------------
