### 2026-01-08 6809062
Update installation instructions to use uv
