### 2026-01-09 3bdb22e
Address JOSS reviewer feedback from @Samreay
- Use consistent absolute imports in __init__.py
- Resolve merge conflict in CONTRIBUTING.md
- Add explicit import statements to documentation examples

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
