+++ jax_supernovae/__init__.py
@@ -1,5 +1,5 @@
 from jax_supernovae.source import SALT3Source, TimeSeriesSource
 from jax_supernovae.data import load_and_process_data
-from . import data
+from jax_supernovae import data
 
 __all__ = ['SALT3Source', 'TimeSeriesSource', 'load_and_process_data', 'data']
