### 2026-01-09 #9441 @samleeney
Thanks for catching these @Samreay! All three points addressed:

1. **Import consistency in `__init__.py`** - Changed the relative import to match the absolute import style used elsewhere

2. **CONTRIBUTING.md conflict** - Resolved, keeping the more comprehensive version

3. **Documentation imports** - Added explicit import statements to the first code block in `bandpass_loading.rst`, `quickstart.rst`, and `model_fluxes.rst` so users can see exactly where each function comes from

And thanks for the kind words about the Best Practices section!
