### 2026-01-28 ba6a82e
Prepare v1.0.0 release for JOSS publication
- Add CITATION.cff for software citation
- Add .zenodo.json for Zenodo archival metadata
- Add CHANGELOG.md documenting version history
- Bump version to 1.0.0

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
