### 2026-01-28 #9441 @samleeney
**Version:** 1.0.0
**Archive DOI:** https://doi.org/10.5281/zenodo.18404558
