### 2026-01-28 21f38c7
Fix version to 1.0.0 and disable auto-increment
- Set version to 1.0.0 for JOSS publication
- Remove auto-increment from publish workflow
- Make tag creation conditional to avoid failures

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
