+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to JAX-bandflux will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.4.0] - 2026-01-28
+## [1.0.0] - 2026-01-28
 
 ### Added
 - CITATION.cff for software citation
