### 2026-01-28 #9441 @JBorrow
Thanks @samleeney. A few copyediting changes for you:

- `≈100×` doesn't render properly. Could you replace this with either `about 100x` or `$\approx 100 \times$`?
- The small sentence fragment 'For example, the reate of expansion of the universe' is odd, and Universe should be capitalised. You could use `(e.g., the rate of expansion of the Universe)` instead of a new sentence here.
- 'Current packages such as SNCosmo' -> 'Existing packages such as SNCosmo'.
- In the bandflux equation use `{\mathrm d}` for the differential element
- Add `# References` at the end so that the references section has a header.

A general question here:

You state in the paper that 'A differentiable approach enables efficient gradient propagation during parameter optimisation and supports large-scale parallel computations on modern hardware such as GPUs'. Is the latter point really a benefit of the differentiability, or is this just another benefit of using JAX? If the latter, this should be re-worded.
