### 2026-01-28 537114a
Address editor copyediting feedback
- Fix ≈100× rendering with LaTeX math
- Fix sentence fragment and capitalize Universe
- Change "Current packages" to "Existing packages"
- Use proper differential notation {\mathrm d}
- Add References header
- Clarify that GPU support is from JAX, not differentiability

Co-Authored-By: Claude Opus 4.5 <noreply@anthropic.com>
