### 2026-01-28 #9441 @samleeney
Thanks @JBorrow for the copyediting feedback. All changes have been addressed in commit [537114a](https://github.com/samleeney/JAX-bandflux/commit/537114a):

- ✅ Fixed `≈100×` rendering → `$\approx 100 \times$`
- ✅ Fixed sentence fragment and capitalised Universe → "(e.g., the rate of expansion of the Universe)"
- ✅ Changed "Current packages" → "Existing packages"
- ✅ Used proper differential notation `{\mathrm d}\lambda`
- ✅ Added `# References` header
- ✅ Clarified that GPU support comes from JAX's compilation model, separate from differentiability
