### 2026-01-28 #9441 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
