### 2026-01-28 #9441 @JBorrow
Thanks for getting back to me so quickly, @samleeney. The final step is a check by the EiC, who may offer some additional copy-editing guidance and could have additional questions. Thanks to our reviewers for their time and effort; your contributions are critical to JOSS' success.
