### 2026-02-03 #9441 @warrickball
Sorry about the wait @samleeney. I've been steadily working through my pre-publication checks and have only a few minor editorial changes that I'll propose via PR hopefully in the next day or two. Once those are merged (presuming you're happy with all the changes), I'll do a final test run of the publication pipeline before publishing.
