### 2026-02-17 #9441 @warrickball
Hi again @samleeney. We're almost there! I apologise but I've introduced an error into the paper with my MR: I forgot to fix the NASA ADS journal macro for the [citation to Kenworthy et al. (2021)](https://github.com/samleeney/JAX-bandflux/blob/e892eae5b98e4d60d4166f0fd2246ba7892fe67c/paper.bib#L25). Could you replace the `\apj` macro with *The Astrophysical Journal*? I could open another PR but it seems a bit excessive.

```diff
@ARTICLE{kenworthy2021salt3,
       author = {{Kenworthy}, W.~D. and {Jones}, D.~O. and {Dai}, M. and {Kessler}, R. and {Scolnic}, D. and {Brout}, D. and {Siebert}, M.~R. and {Pierel}, J.~D.~R. and {Dettman}, K.~G. and {Dimitriadis}, G. and {Foley}, R.~J. and {Jha}, S.~W. and {Pan}, Y.-C. and {Riess}, A. and {Rodney}, S. and {Rojas-Bravo}, C.},
        title = {SALT3: An Improved {Type Ia} Supernova Model for Measuring Cosmic Distances},
-      journal = {\apj},
+      journal = {The Astrophysical Journal},
     keywords = {1146, 1728, Astrophysics - Cosmology and Nongalactic Astrophysics, Astrophysics - Instrumentation and Methods for Astrophysics},

```
