### 2026-03-03 #9441 @samleeney
Done — I've replaced the `\apj` macro with *The Astrophysical Journal* in the Kenworthy et al. (2021) entry as requested, and also caught the same issue in the Pierel et al. (2022) entry so fixed that too: https://github.com/samleeney/JAX-bandflux/commit/398bc7e
