### 2026-03-03 #9441 @editorialbot
:wave: @openjournals/aass-eics, this paper is ready to be accepted and published.
