### 2026-03-03 #9441 @warrickball
Congratulations @samleeney, your paper has been published in JOSS! My apologies once again for introducing those errors in the journal titles.
